DeepFilterNet: split 2980-line monolith into 5 focused files
Break DeepFilterNetModel.swift into well-organized files by concern:
- DeepFilterNetModel.swift (448 lines): error types, config, model class, loading, public API
- DeepFilterNetStreamer.swift (897 lines): streaming pipeline as top-level class
- DeepFilterNetForward.swift (777 lines): network forward pass (V1/V2/V3)
- DeepFilterNetLayers.swift (455 lines): conv2d, convTranspose, batchNorm, GRU, linear
- DeepFilterNetDSP.swift (433 lines): feature normalization, ERB, weight cache builders

Also adds benchmark scripts (speed, accuracy, latency, memory) used to verify
zero regression: all accuracy metrics are bit-identical before/after.
