+++ Sources/MLXAudioSTS/Models/DeepFilterNet/DeepFilterNetModel.swift
@@ -76,6 +76,8 @@ public struct DeepFilterNetStreamingChunk: @unchecked Sendable {
     public let isLastChunk: Bool
 }
 
+// MARK: - Model
+
 public final class DeepFilterNetModel: STSModel {
     public static let defaultRepo = "iky1e/DeepFilterNet3-MLX"
 
@@ -86,28 +88,32 @@ public final class DeepFilterNetModel: STSModel {
     public var supportsStreaming: Bool { !isV1 }
     public var sampleRate: Int { config.sampleRate }
 
-    private let weights: [String: MLXArray]
-    private let erbFB: MLXArray
-    private let erbInvFB: MLXArray
-    private let erbBandWidths: [Int]
-    private let vorbisWindow: MLXArray
-    private let wnorm: Float
-    private let normAlphaValue: Float
-    private let inferenceDType: DType
-    private let bnScale: [String: MLXArray]
-    private let bnBias: [String: MLXArray]
-    private let conv2dWeightsOHWI: [String: MLXArray]
-    private let convTransposeDenseWeights: [String: MLXArray]
-    private let convTransposeGroupWeights: [String: [MLXArray]]
-    private let gruTransposedWeights: [String: MLXArray]
-    private struct V1GroupedLinearPack {
+    // Internal visibility for cross-file access by extensions and DeepFilterNetStreamer.
+    let weights: [String: MLXArray]
+    let erbFB: MLXArray
+    let erbInvFB: MLXArray
+    let erbBandWidths: [Int]
+    let vorbisWindowArray: MLXArray
+    let wnorm: Float
+    let normAlphaValue: Float
+    let inferenceDType: DType
+    let bnScale: [String: MLXArray]
+    let bnBias: [String: MLXArray]
+    let conv2dWeightsOHWI: [String: MLXArray]
+    let convTransposeDenseWeights: [String: MLXArray]
+    let convTransposeGroupWeights: [String: [MLXArray]]
+    let gruTransposedWeights: [String: MLXArray]
+    let j: MLXArray = MLXArray(real: Float(0.0), imaginary: Float(1.0))
+
+    struct V1GroupedLinearPack {
         let weightGIO: MLXArray  // [G, I, O]
         let biasGO: MLXArray  // [G, O]
         let groups: Int
         let inputPerGroup: Int
         let outputPerGroup: Int
     }
-    private struct V1GroupedGRULayerPack {
+
+    struct V1GroupedGRULayerPack {
         let weightIHGI3H: MLXArray  // [G, I, 3H]
         let weightHHGH3H: MLXArray  // [G, H, 3H]
         let biasIHG3H: MLXArray  // [G, 3H]
@@ -115,13 +121,14 @@ public final class DeepFilterNetModel: STSModel {
         let inputPerGroup: Int
         let hiddenPerGroup: Int
     }
-    private struct V1GroupedGRUPack {
+
+    struct V1GroupedGRUPack {
         let groups: Int
         let layers: [V1GroupedGRULayerPack]
     }
-    private let v1GroupedLinearPacks: [String: V1GroupedLinearPack]
-    private let v1GroupedGRUPacks: [String: V1GroupedGRUPack]
-    private let j: MLXArray = MLXArray(real: Float(0.0), imaginary: Float(1.0))
+
+    let v1GroupedLinearPacks: [String: V1GroupedLinearPack]
+    let v1GroupedGRUPacks: [String: V1GroupedGRUPack]
 
     private init(
         config: DeepFilterNetConfig,
@@ -149,7 +156,7 @@ public final class DeepFilterNetModel: STSModel {
                 minNbFreqs: max(1, config.minNbErbFreqs)
             )
         }
-        self.vorbisWindow = Self.vorbisWindow(size: config.fftSize)
+        self.vorbisWindowArray = Self.vorbisWindow(size: config.fftSize)
         self.wnorm = 1.0 / Float(config.fftSize * config.fftSize) * Float(2 * config.hopSize)
         self.normAlphaValue = Self.computeNormAlpha(hopSize: config.hopSize, sampleRate: config.sampleRate)
         self.inferenceDType =
@@ -256,7 +263,7 @@ public final class DeepFilterNetModel: STSModel {
             fftLen: config.fftSize,
             hopLength: config.hopSize,
             winLen: config.fftSize,
-            window: vorbisWindow,
+            window: vorbisWindowArray,
             center: false
         )
         let spec = specComplex * MLXArray(wnorm)
@@ -297,8 +304,7 @@ public final class DeepFilterNetModel: STSModel {
             )
         }
         let specEnhanced = forwardOut.0
-        // Keep shape handling robust across MLX indexing semantics:
-        // expected post-squeeze shape is [T, F, 2] before ISTFT layout conversion.
+
         var enhTF2 = specEnhanced
             .squeezed(axis: 0)
             .squeezed(axis: 0)
@@ -311,8 +317,6 @@ public final class DeepFilterNetModel: STSModel {
         var enhReal2D = enh.realPart().squeezed()
         var enhImag2D = enh.imaginaryPart().squeezed()
         if enhReal2D.ndim != 2 || enhImag2D.ndim != 2 {
-            // Fallback for unexpected singleton-preserving index behavior.
-            // Expected semantic shape is [T, F], so collapse leading dims to recover it.
             let t = spec.shape[2]
             let f = config.freqBins
             enhReal2D = enhReal2D.reshaped([t, f])
@@ -327,7 +331,7 @@ public final class DeepFilterNetModel: STSModel {
             fftLen: config.fftSize,
             hopLength: config.hopSize,
             winLen: config.fftSize,
-            window: vorbisWindow,
+            window: vorbisWindowArray,
             center: false,
             audioLength: origLen + config.hopSize + config.fftSize
         )
@@ -365,7 +369,6 @@ public final class DeepFilterNetModel: STSModel {
         }
 
         let streamer = createStreamer(config: config)
-        // Default to true low-latency chunking: one hop (10ms at 48kHz).
         let frameChunk = max(self.config.hopSize, chunkSamples ?? self.config.hopSize)
         var outputChunks = [MLXArray]()
         outputChunks.reserveCapacity(max(1, samples.shape[0] / frameChunk))
@@ -405,7 +408,6 @@ public final class DeepFilterNetModel: STSModel {
                 }
                 let samples = audioInput.asType(.float32)
                 let streamer = createStreamer(config: config)
-                // Default to true low-latency chunking: one hop (10ms at 48kHz).
                 let frameChunk = max(self.config.hopSize, chunkSamples ?? self.config.hopSize)
 
                 var chunkIndex = 0
@@ -442,2539 +444,5 @@ public final class DeepFilterNetModel: STSModel {
                 continuation.finish(throwing: error)
             }
         }
-    }
-
-    // MARK: - Streaming
-
-    public final class DeepFilterNetStreamer {
-        private struct StreamGRULayer {
-            let wihT: MLXArray
-            let whhT: MLXArray
-            let bih: MLXArray
-            let bhh: MLXArray
-        }
-
-        private struct StreamGRU {
-            let linearInWeight: MLXArray
-            let layers: [StreamGRULayer]
-            let linearOutWeight: MLXArray?
-        }
-
-        private struct StaticRing {
-            private(set) var values: [MLXArray]
-            private(set) var totalWritten: Int = 0
-
-            var capacity: Int { values.count }
-            var count: Int { min(totalWritten, capacity) }
-            var oldestAbsoluteIndex: Int { max(0, totalWritten - capacity) }
-
-            init(capacity: Int, initial: MLXArray) {
-                precondition(capacity > 0, "StaticRing capacity must be > 0")
-                values = Array(repeating: initial, count: capacity)
-            }
-
-            mutating func reset() {
-                totalWritten = 0
-            }
-
-            mutating func push(_ value: MLXArray) {
-                values[totalWritten % capacity] = value
-                totalWritten += 1
-            }
-
-            func get(absoluteIndex: Int) -> MLXArray? {
-                guard absoluteIndex >= oldestAbsoluteIndex, absoluteIndex < totalWritten else {
-                    return nil
-                }
-                return values[absoluteIndex % capacity]
-            }
-
-            func orderedLast(_ n: Int) -> [MLXArray] {
-                let k = min(max(0, n), totalWritten)
-                guard k > 0 else { return [] }
-                let start = totalWritten - k
-                return (start..<(start + k)).compactMap { get(absoluteIndex: $0) }
-            }
-        }
-
-        private let model: DeepFilterNetModel
-        public let config: DeepFilterNetStreamingConfig
-
-        private let fftSize: Int
-        private let hopSize: Int
-        private let freqBins: Int
-        private let nbDf: Int
-        private let nbErb: Int
-        private let dfOrder: Int
-        private let dfLookahead: Int
-        private let convLookahead: Int
-
-        private let alphaArray: MLXArray
-        private let oneMinusAlphaArray: MLXArray
-        private let fftScaleArray: MLXArray
-        private let vorbisWindow: MLXArray
-        private let wnormArray: MLXArray
-        private let inferenceDType: DType
-        private let epsEnergy = MLXArray(Float(1e-10))
-        private let epsNorm = MLXArray(Float(1e-12))
-        private let tenArray = MLXArray(Float(10.0))
-        private let fortyArray = MLXArray(Float(40.0))
-        private let zeroSpecFrame: MLXArray
-        private let zeroSpecLowFrame: MLXArray
-        private let zeroMaskFrame: MLXArray
-        private let zeroEncErbFrame: MLXArray
-        private let zeroEncDfFrame: MLXArray
-        private let zeroDfConvpFrame: MLXArray
-        private let specRingCapacity: Int
-        private let dfConvpKernelSizeT: Int
-        private let dfSpecLeft: Int
-        private let analysisMemCount: Int
-        private let synthMemCount: Int
-        private let erbFBFrame: MLXArray?
-        private let lsnrWeight: MLXArray
-        private let lsnrBias: MLXArray
-        private let lsnrScale: MLXArray
-        private let lsnrOffset: MLXArray
-        private let encDfFcEmbWeight: MLXArray
-        private let dfDecSkipWeight: MLXArray?
-        private let dfDecOutWeight: MLXArray
-        private let encEmbGRU: StreamGRU
-        private let erbDecEmbGRU: StreamGRU
-        private let dfDecGRU: StreamGRU
-
-        private var pendingSamples = MLXArray.zeros([0], type: Float.self)
-        private var analysisMem: MLXArray
-        private var synthMem: MLXArray
-        private var erbState: MLXArray
-        private var dfState: MLXArray
-
-        private var specRing: StaticRing
-        private var encErbHistory: MLXArray
-        private var encDfHistory: MLXArray
-        private var dfConvpHistory: MLXArray
-        private var dfSpecHistory: MLXArray
-        private var dfSpecHistoryInitialized = false
-
-        private var encEmbState: [MLXArray]?
-        private var erbDecState: [MLXArray]?
-        private var dfDecState: [MLXArray]?
-
-        private var delayDropped = 0
-        private var hopsSinceMaterialize = 0
-        private let enableProfiling: Bool
-        private var profHopCount = 0
-        private var profAnalysisSeconds = 0.0
-        private var profFeaturesSeconds = 0.0
-        private var profInferSeconds = 0.0
-        private var profInferEncodeSeconds = 0.0
-        private var profInferEmbSeconds = 0.0
-        private var profInferErbSeconds = 0.0
-        private var profInferDfSeconds = 0.0
-        private var profSynthesisSeconds = 0.0
-        private var profMaterializeSeconds = 0.0
-        private let profilingForceEvalPerStage: Bool
-
-        public init(model: DeepFilterNetModel, config: DeepFilterNetStreamingConfig = DeepFilterNetStreamingConfig()) {
-            self.model = model
-            self.config = config
-            self.enableProfiling = config.enableProfiling
-            self.profilingForceEvalPerStage = config.profilingForceEvalPerStage
-
-            self.fftSize = model.config.fftSize
-            self.hopSize = model.config.hopSize
-            self.freqBins = model.config.freqBins
-            self.nbDf = model.config.nbDf
-            self.nbErb = model.config.nbErb
-            self.dfOrder = model.config.dfOrder
-            self.dfLookahead = model.config.dfLookahead
-            self.convLookahead = model.config.convLookahead
-
-            let alpha = model.normAlpha()
-            self.alphaArray = MLXArray(alpha)
-            self.oneMinusAlphaArray = MLXArray(Float(1.0) - alpha)
-            self.fftScaleArray = MLXArray(Float(model.config.fftSize))
-            self.vorbisWindow = model.vorbisWindow.asType(.float32)
-            self.wnormArray = MLXArray(model.wnorm)
-            self.inferenceDType = model.inferenceDType
-            self.analysisMemCount = max(0, model.config.fftSize - model.config.hopSize)
-            self.synthMemCount = max(0, model.config.fftSize - model.config.hopSize)
-            self.zeroSpecFrame = MLXArray.zeros([model.config.freqBins, 2], type: Float.self)
-            self.zeroSpecLowFrame = zeroSpecFrame[0..<model.config.nbDf, 0...]
-            self.zeroMaskFrame = MLXArray.zeros([1, 1, 1, model.config.nbErb], type: Float.self)
-            self.zeroEncErbFrame = MLXArray.zeros([1, 1, 1, model.config.nbErb], type: Float.self).asType(model.inferenceDType)
-            self.zeroEncDfFrame = MLXArray.zeros([1, 2, 1, model.config.nbDf], type: Float.self).asType(model.inferenceDType)
-            self.zeroDfConvpFrame = MLXArray.zeros([1, model.config.convCh, 1, model.config.nbDf], type: Float.self).asType(model.inferenceDType)
-            self.dfConvpKernelSizeT = max(1, model.config.dfPathwayKernelSizeT)
-            let leftHistory = max(0, model.config.dfOrder - model.config.dfLookahead - 1)
-            self.dfSpecLeft = leftHistory
-            self.specRingCapacity = max(8, leftHistory + model.config.convLookahead + model.config.dfLookahead + 4)
-            if model.erbFB.shape.count == 2,
-               model.erbFB.shape[0] == model.config.freqBins,
-               model.erbFB.shape[1] == model.config.nbErb
-            {
-                self.erbFBFrame = model.erbFB.asType(.float32)
-            } else {
-                self.erbFBFrame = nil
-            }
-            self.lsnrWeight = (try? model.w("enc.lsnr_fc.0.weight")) ?? MLXArray.zeros([1, model.config.embHiddenDim], type: Float.self)
-            self.lsnrBias = (try? model.w("enc.lsnr_fc.0.bias")) ?? MLXArray.zeros([1], type: Float.self)
-            self.lsnrScale = MLXArray(Float(model.config.lsnrMax - model.config.lsnrMin))
-            self.lsnrOffset = MLXArray(Float(model.config.lsnrMin))
-            self.encDfFcEmbWeight = Self.requireWeight(model, key: "enc.df_fc_emb.0.weight")
-            self.dfDecSkipWeight = model.weights["df_dec.df_skip.weight"]
-            self.dfDecOutWeight = Self.requireWeight(model, key: "df_dec.df_out.0.weight")
-            self.encEmbGRU = Self.buildStreamGRU(model: model, prefix: "enc.emb_gru")
-            self.erbDecEmbGRU = Self.buildStreamGRU(model: model, prefix: "erb_dec.emb_gru")
-            self.dfDecGRU = Self.buildStreamGRU(model: model, prefix: "df_dec.df_gru")
-
-            self.analysisMem = MLXArray.zeros([analysisMemCount], type: Float.self)
-            self.synthMem = MLXArray.zeros([synthMemCount], type: Float.self)
-            self.erbState = MLXArray(Self.linspace(start: -60.0, end: -90.0, count: model.config.nbErb))
-            self.dfState = MLXArray(Self.linspace(start: 0.001, end: 0.0001, count: model.config.nbDf))
-            self.specRing = StaticRing(capacity: specRingCapacity, initial: zeroSpecFrame)
-            self.encErbHistory = MLX.repeated(zeroEncErbFrame, count: 3, axis: 2)
-            self.encDfHistory = MLX.repeated(zeroEncDfFrame, count: 3, axis: 2)
-            self.dfConvpHistory = MLX.repeated(zeroDfConvpFrame, count: dfConvpKernelSizeT, axis: 2)
-            self.dfSpecHistory = MLX.repeated(zeroSpecLowFrame.expandedDimensions(axis: 0), count: model.config.dfOrder, axis: 0)
-        }
-
-        private static func requireWeight(_ model: DeepFilterNetModel, key: String) -> MLXArray {
-            guard let weight = model.weights[key] else {
-                preconditionFailure("Missing required DeepFilterNet weight: \(key)")
-            }
-            return weight
-        }
-
-        private static func buildStreamGRU(model: DeepFilterNetModel, prefix: String) -> StreamGRU {
-            let linearIn = requireWeight(model, key: "\(prefix).linear_in.0.weight")
-            let linearOut = model.weights["\(prefix).linear_out.0.weight"]
-
-            var layers = [StreamGRULayer]()
-            var layer = 0
-            while model.weights["\(prefix).gru.weight_ih_l\(layer)"] != nil {
-                let wihKey = "\(prefix).gru.weight_ih_l\(layer)"
-                let whhKey = "\(prefix).gru.weight_hh_l\(layer)"
-                let wihT = model.gruTransposedWeights[wihKey] ?? requireWeight(model, key: wihKey).transposed()
-                let whhT = model.gruTransposedWeights[whhKey] ?? requireWeight(model, key: whhKey).transposed()
-                let bih = requireWeight(model, key: "\(prefix).gru.bias_ih_l\(layer)")
-                let bhh = requireWeight(model, key: "\(prefix).gru.bias_hh_l\(layer)")
-                layers.append(StreamGRULayer(wihT: wihT, whhT: whhT, bih: bih, bhh: bhh))
-                layer += 1
-            }
-
-            return StreamGRU(linearInWeight: linearIn, layers: layers, linearOutWeight: linearOut)
-        }
-
-        public func reset() {
-            pendingSamples = MLXArray.zeros([0], type: Float.self)
-            analysisMem = MLXArray.zeros([analysisMemCount], type: Float.self)
-            synthMem = MLXArray.zeros([synthMemCount], type: Float.self)
-            erbState = MLXArray(Self.linspace(start: -60.0, end: -90.0, count: nbErb))
-            dfState = MLXArray(Self.linspace(start: 0.001, end: 0.0001, count: nbDf))
-            specRing.reset()
-            encErbHistory = MLX.repeated(zeroEncErbFrame, count: 3, axis: 2)
-            encDfHistory = MLX.repeated(zeroEncDfFrame, count: 3, axis: 2)
-            dfConvpHistory = MLX.repeated(zeroDfConvpFrame, count: dfConvpKernelSizeT, axis: 2)
-            dfSpecHistory = MLX.repeated(zeroSpecLowFrame.expandedDimensions(axis: 0), count: dfOrder, axis: 0)
-            dfSpecHistoryInitialized = false
-            encEmbState = nil
-            erbDecState = nil
-            dfDecState = nil
-            delayDropped = 0
-            hopsSinceMaterialize = 0
-            profHopCount = 0
-            profAnalysisSeconds = 0.0
-            profFeaturesSeconds = 0.0
-            profInferSeconds = 0.0
-            profInferEncodeSeconds = 0.0
-            profInferEmbSeconds = 0.0
-            profInferErbSeconds = 0.0
-            profInferDfSeconds = 0.0
-            profSynthesisSeconds = 0.0
-            profMaterializeSeconds = 0.0
-        }
-
-        public func processChunk(_ chunk: MLXArray, isLast: Bool = false) throws -> MLXArray {
-            guard chunk.ndim == 1 else {
-                throw DeepFilterNetError.invalidAudioShape(chunk.shape)
-            }
-            let chunkF32 = chunk.asType(.float32)
-            if !isLast, pendingSamples.shape[0] == 0, chunkF32.shape[0] == hopSize {
-                if var out = try processHop(chunkF32) {
-                    if config.compensateDelay {
-                        let totalDelay = fftSize - hopSize
-                        if delayDropped < totalDelay {
-                            let toDrop = min(totalDelay - delayDropped, out.shape[0])
-                            if toDrop > 0 {
-                                out = out[toDrop..<out.shape[0]]
-                                delayDropped += toDrop
-                            }
-                        }
-                    }
-                    return out
-                }
-                return MLXArray.zeros([0], type: Float.self)
-            }
-            if chunkF32.shape[0] > 0 {
-                if pendingSamples.shape[0] == 0 {
-                    pendingSamples = chunkF32
-                } else {
-                    pendingSamples = MLX.concatenated([pendingSamples, chunkF32], axis: 0)
-                }
-            }
-
-            var outs = [MLXArray]()
-            while pendingSamples.shape[0] >= hopSize {
-                let hop = pendingSamples[0..<hopSize]
-                pendingSamples = pendingSamples[hopSize..<pendingSamples.shape[0]]
-                if let out = try processHop(hop) {
-                    outs.append(out)
-                }
-            }
-
-            if isLast {
-                if config.padEndFrames > 0 {
-                    let pad = MLXArray.zeros([config.padEndFrames * hopSize], type: Float.self)
-                    if pendingSamples.shape[0] == 0 {
-                        pendingSamples = pad
-                    } else {
-                        pendingSamples = MLX.concatenated([pendingSamples, pad], axis: 0)
-                    }
-                }
-                while pendingSamples.shape[0] >= hopSize {
-                    let hop = pendingSamples[0..<hopSize]
-                    pendingSamples = pendingSamples[hopSize..<pendingSamples.shape[0]]
-                    if let out = try processHop(hop) {
-                        outs.append(out)
-                    }
-                }
-            }
-
-            var y: MLXArray
-            if outs.isEmpty {
-                y = MLXArray.zeros([0], type: Float.self)
-            } else if outs.count == 1, let first = outs.first {
-                y = first
-            } else {
-                y = MLX.concatenated(outs, axis: 0)
-            }
-
-            if config.compensateDelay {
-                let totalDelay = fftSize - hopSize
-                if delayDropped < totalDelay {
-                    let toDrop = min(totalDelay - delayDropped, y.shape[0])
-                    if toDrop > 0 {
-                        y = y[toDrop..<y.shape[0]]
-                        delayDropped += toDrop
-                    }
-                }
-            }
-
-            return y
-        }
-
-        public func processChunk(_ chunk: [Float], isLast: Bool = false) throws -> [Float] {
-            guard !chunk.isEmpty || isLast else { return [] }
-            let y = try processChunk(MLXArray(chunk), isLast: isLast)
-            if y.shape[0] == 0 {
-                return []
-            }
-            return y.asArray(Float.self)
-        }
-
-        public func flush() throws -> [Float] {
-            try processChunk([], isLast: true)
-        }
-
-        public func flushMLX() throws -> MLXArray {
-            try processChunk(MLXArray.zeros([0], type: Float.self), isLast: true)
-        }
-
-        public func profilingSummary() -> String? {
-            guard enableProfiling else { return nil }
-            let hops = max(profHopCount, 1)
-            let total = profAnalysisSeconds + profFeaturesSeconds + profInferSeconds + profSynthesisSeconds + profMaterializeSeconds
-            let perHopMs = (total / Double(hops)) * 1000.0
-            func pct(_ v: Double) -> Double {
-                guard total > 0 else { return 0.0 }
-                return (v / total) * 100.0
-            }
-            return String(
-                format:
-                    """
-                    Stream profile: hops=%d total=%.3fs perHop=%.3fms
-                      analysis:    %.3fs (%.1f%%)
-                      features:    %.3fs (%.1f%%)
-                      infer:       %.3fs (%.1f%%)
-                        infer.enc: %.3fs (%.1f%% infer)
-                        infer.emb: %.3fs (%.1f%% infer)
-                        infer.erb: %.3fs (%.1f%% infer)
-                        infer.df:  %.3fs (%.1f%% infer)
-                      synthesis:   %.3fs (%.1f%%)
-                      materialize: %.3fs (%.1f%%)
-                    """,
-                profHopCount,
-                total,
-                perHopMs,
-                profAnalysisSeconds, pct(profAnalysisSeconds),
-                profFeaturesSeconds, pct(profFeaturesSeconds),
-                profInferSeconds, pct(profInferSeconds),
-                profInferEncodeSeconds, profInferSeconds > 0 ? (profInferEncodeSeconds / profInferSeconds) * 100.0 : 0.0,
-                profInferEmbSeconds, profInferSeconds > 0 ? (profInferEmbSeconds / profInferSeconds) * 100.0 : 0.0,
-                profInferErbSeconds, profInferSeconds > 0 ? (profInferErbSeconds / profInferSeconds) * 100.0 : 0.0,
-                profInferDfSeconds, profInferSeconds > 0 ? (profInferDfSeconds / profInferSeconds) * 100.0 : 0.0,
-                profSynthesisSeconds, pct(profSynthesisSeconds),
-                profMaterializeSeconds, pct(profMaterializeSeconds)
-            )
-        }
-
-        private func processHop(_ hopTD: MLXArray) throws -> MLXArray? {
-            let tAnalysis0 = enableProfiling ? CFAbsoluteTimeGetCurrent() : 0
-            let spec = analysisFrame(hopTD)
-            if enableProfiling, profilingForceEvalPerStage {
-                eval(spec)
-            }
-            if enableProfiling {
-                profAnalysisSeconds += CFAbsoluteTimeGetCurrent() - tAnalysis0
-            }
-
-            let tFeatures0 = enableProfiling ? CFAbsoluteTimeGetCurrent() : 0
-            let (featErb, featDf) = featuresFrame(spec)
-            if enableProfiling, profilingForceEvalPerStage {
-                eval(featErb, featDf)
-            }
-            if enableProfiling {
-                profFeaturesSeconds += CFAbsoluteTimeGetCurrent() - tFeatures0
-            }
-            specRing.push(spec)
-            encErbHistory = appendHistoryFrame(encErbHistory, frame: featErb.asType(inferenceDType))
-            encDfHistory = appendHistoryFrame(encDfHistory, frame: featDf.asType(inferenceDType))
-
-            if specRing.totalWritten <= convLookahead {
-                return nil
-            }
-            let targetFrameIndex = specRing.totalWritten - 1 - convLookahead
-            guard let specT = specRing.get(absoluteIndex: targetFrameIndex) else {
-                return nil
-            }
-            updateDfSpecHistory(targetFrameIndex: targetFrameIndex)
-            let tInfer0 = enableProfiling ? CFAbsoluteTimeGetCurrent() : 0
-            let specEnhanced = try inferFrame(spec: specT, targetFrameIndex: targetFrameIndex)
-            if enableProfiling, profilingForceEvalPerStage {
-                eval(specEnhanced)
-            }
-            if enableProfiling {
-                profInferSeconds += CFAbsoluteTimeGetCurrent() - tInfer0
-            }
-
-            let tSynth0 = enableProfiling ? CFAbsoluteTimeGetCurrent() : 0
-            let out = synthesisFrame(specEnhanced.asType(.float32))
-            if enableProfiling, profilingForceEvalPerStage {
-                eval(out)
-            }
-            if enableProfiling {
-                profSynthesisSeconds += CFAbsoluteTimeGetCurrent() - tSynth0
-            }
-            hopsSinceMaterialize += 1
-            if config.materializeEveryHops > 0, hopsSinceMaterialize >= config.materializeEveryHops {
-                let tMat0 = enableProfiling ? CFAbsoluteTimeGetCurrent() : 0
-                materializeStreamingState(output: out)
-                if enableProfiling {
-                    profMaterializeSeconds += CFAbsoluteTimeGetCurrent() - tMat0
-                }
-                hopsSinceMaterialize = 0
-            }
-            if enableProfiling {
-                profHopCount += 1
-            }
-            return out
-        }
-
-        private func analysisFrame(_ hopTD: MLXArray) -> MLXArray {
-            let frame = analysisMemCount > 0
-                ? MLX.concatenated([analysisMem, hopTD], axis: 0)
-                : hopTD
-            let frameWin = frame * vorbisWindow
-            let specComplex = MLXFFT.rfft(frameWin, axis: 0) * wnormArray
-            let spec = MLX.stacked([specComplex.realPart(), specComplex.imaginaryPart()], axis: -1)
-            updateAnalysisMemory(with: hopTD)
-            return spec
-        }
-
-        private func synthesisFrame(_ specNorm: MLXArray) -> MLXArray {
-            let complex = specNorm[0..., 0] + model.j * specNorm[0..., 1]
-            var time = MLXFFT.irfft(complex, axis: 0)
-            time = time * fftScaleArray
-            time = time * vorbisWindow
-
-            let out = time[0..<hopSize] + synthMem[0..<hopSize]
-            updateSynthesisMemory(with: time)
-            return out
-        }
-
-        private func updateAnalysisMemory(with hop: MLXArray) {
-            guard analysisMemCount > 0 else { return }
-            if analysisMemCount > hopSize {
-                let split = analysisMemCount - hopSize
-                let rotated = MLX.concatenated([
-                    analysisMem[hopSize..<analysisMemCount],
-                    analysisMem[0..<hopSize],
-                ], axis: 0)
-                analysisMem = MLX.concatenated([rotated[0..<split], hop], axis: 0)
-            } else {
-                analysisMem = hop[(hopSize - analysisMemCount)..<hopSize]
-            }
-        }
-
-        private func updateSynthesisMemory(with time: MLXArray) {
-            guard synthMemCount > 0 else { return }
-            let xSecond = time[hopSize..<fftSize]
-            if synthMemCount > hopSize {
-                let split = synthMemCount - hopSize
-                let rotated = MLX.concatenated([
-                    synthMem[hopSize..<synthMemCount],
-                    synthMem[0..<hopSize],
-                ], axis: 0)
-                let sFirst = rotated[0..<split] + xSecond[0..<split]
-                let sSecond = xSecond[split..<(split + hopSize)]
-                synthMem = MLX.concatenated([sFirst, sSecond], axis: 0)
-            } else {
-                synthMem = xSecond[0..<synthMemCount]
-            }
-        }
-
-        private func featuresFrame(_ spec: MLXArray) -> (MLXArray, MLXArray) {
-            let re = spec[0..., 0]
-            let im = spec[0..., 1]
-            let magSq = re.square() + im.square()
-
-            let erb: MLXArray
-            if let erbFBFrame {
-                erb = MLX.matmul(magSq.expandedDimensions(axis: 0), erbFBFrame).squeezed()
-            } else {
-                var erbBands = [MLXArray]()
-                erbBands.reserveCapacity(nbErb)
-                var start = 0
-                for width in model.erbBandWidths {
-                    let stop = min(start + width, freqBins)
-                    if stop > start {
-                        erbBands.append(MLX.mean(magSq[start..<stop], axis: 0))
-                    } else {
-                        erbBands.append(MLXArray.zeros([1], type: Float.self).squeezed())
-                    }
-                    start = stop
-                }
-                erb = MLX.stacked(erbBands, axis: 0)
-            }
-            let erbDB = tenArray * MLX.log10(erb + epsEnergy)
-            erbState = erbDB * oneMinusAlphaArray + erbState * alphaArray
-            let featErb = (erbDB - erbState) / fortyArray
-
-            let dfRe = re[0..<nbDf]
-            let dfIm = im[0..<nbDf]
-            let mag = MLX.sqrt(dfRe.square() + dfIm.square())
-            dfState = mag * oneMinusAlphaArray + dfState * alphaArray
-            let denom = MLX.sqrt(MLX.maximum(dfState, epsNorm))
-            let featDfRe = dfRe / denom
-            let featDfIm = dfIm / denom
-
-            let featErbMX = featErb
-                .expandedDimensions(axis: 0)
-                .expandedDimensions(axis: 0)
-                .expandedDimensions(axis: 0)
-            var featDfMX = MLX.stacked([featDfRe, featDfIm], axis: -1)
-                .expandedDimensions(axis: 0)
-                .expandedDimensions(axis: 0)
-            featDfMX = featDfMX.transposed(0, 3, 1, 2)
-            return (featErbMX, featDfMX)
-        }
-
-        private func inferFrame(
-            spec: MLXArray,
-            targetFrameIndex: Int
-        ) throws -> MLXArray {
-            let tEncode0 = enableProfiling ? CFAbsoluteTimeGetCurrent() : 0
-            let specMX = spec
-                .expandedDimensions(axis: 0)
-                .expandedDimensions(axis: 0)
-                .expandedDimensions(axis: 0)
-                .asType(inferenceDType)
-
-            let encErbSeq = encErbHistory
-            let encDfSeq = encDfHistory
-
-            let e0 = try applyConvLast(input: encErbSeq, prefix: "enc.erb_conv0", main: 1, pointwise: nil, bn: 2, fstride: 1)
-            let e1 = try applyConvLast(input: e0, prefix: "enc.erb_conv1", main: 0, pointwise: 1, bn: 2, fstride: 2)
-            let e2 = try applyConvLast(input: e1, prefix: "enc.erb_conv2", main: 0, pointwise: 1, bn: 2, fstride: 2)
-            let e3 = try applyConvLast(input: e2, prefix: "enc.erb_conv3", main: 0, pointwise: 1, bn: 2, fstride: 1)
-
-            let c0 = try applyConvLast(input: encDfSeq, prefix: "enc.df_conv0", main: 1, pointwise: 2, bn: 3, fstride: 1)
-            let c1 = try applyConvLast(input: c0, prefix: "enc.df_conv1", main: 0, pointwise: 1, bn: 2, fstride: 2)
-
-            var cemb = c1.transposed(0, 2, 3, 1).reshaped([1, 1, -1])
-            cemb = relu(model.groupedLinear(cemb, weight: encDfFcEmbWeight))
-
-            var emb = e3.transposed(0, 2, 3, 1).reshaped([1, 1, -1])
-            emb = model.config.encConcat ? MLX.concatenated([emb, cemb], axis: -1) : (emb + cemb)
-            if enableProfiling, profilingForceEvalPerStage {
-                eval(e3, c1, emb)
-            }
-            if enableProfiling {
-                profInferEncodeSeconds += CFAbsoluteTimeGetCurrent() - tEncode0
-            }
-
-            let tEmb0 = enableProfiling ? CFAbsoluteTimeGetCurrent() : 0
-            emb = squeezedGRUStep(
-                emb,
-                gru: encEmbGRU,
-                hiddenSize: model.config.embHiddenDim,
-                state: &encEmbState
-            )
-
-            let applyGains: Bool
-            let applyGainZeros: Bool
-            let applyDf: Bool
-            if config.enableStageSkipping {
-                let lsnr = sigmoid(model.linear(emb, weight: lsnrWeight, bias: lsnrBias)) * lsnrScale + lsnrOffset
-                let lsnrValue = lsnr.asArray(Float.self).first ?? Float(model.config.lsnrMin)
-                (applyGains, applyGainZeros, applyDf) = applyStages(lsnr: lsnrValue)
-            } else {
-                (applyGains, applyGainZeros, applyDf) = (true, false, true)
-            }
-            if enableProfiling, profilingForceEvalPerStage {
-                eval(emb)
-            }
-            if enableProfiling {
-                profInferEmbSeconds += CFAbsoluteTimeGetCurrent() - tEmb0
-            }
-
-            let mask: MLXArray
-            let tErb0 = enableProfiling ? CFAbsoluteTimeGetCurrent() : 0
-            if applyGains {
-                mask = try erbDecoderStep(emb: emb, e3: e3, e2: e2, e1: e1, e0: e0)
-            } else if applyGainZeros {
-                mask = zeroMaskFrame.asType(inferenceDType)
-            } else {
-                return specMX[0, 0, 0, 0..., 0...]
-            }
-            let specMasked = applyMask(spec: specMX, mask: mask)
-            if enableProfiling, profilingForceEvalPerStage {
-                eval(mask, specMasked)
-            }
-            if enableProfiling {
-                profInferErbSeconds += CFAbsoluteTimeGetCurrent() - tErb0
-            }
-            if !applyDf {
-                return specMasked[0, 0, 0, 0..., 0...]
-            }
-
-            let tDf0 = enableProfiling ? CFAbsoluteTimeGetCurrent() : 0
-            var dfCoefs = try dfDecoderStep(emb: emb, c0: c0)
-            dfCoefs = dfCoefs.reshaped([1, 1, nbDf, dfOrder, 2]).transposed(0, 3, 1, 2, 4)
-
-            let specEnhanced = try deepFilterAssign(
-                spec: specMX,
-                specMasked: specMasked,
-                dfCoefs: dfCoefs,
-                targetFrameIndex: targetFrameIndex
-            )
-            if enableProfiling, profilingForceEvalPerStage {
-                eval(dfCoefs, specEnhanced)
-            }
-            if enableProfiling {
-                profInferDfSeconds += CFAbsoluteTimeGetCurrent() - tDf0
-            }
-            return specEnhanced[0, 0, 0, 0..., 0...]
-        }
-
-        private func applyStages(lsnr: Float) -> (Bool, Bool, Bool) {
-            if lsnr < config.minDbThresh {
-                // Only noise detected: apply zero ERB mask and skip DF.
-                return (false, true, false)
-            }
-            if lsnr > config.maxDbErbThresh {
-                // Clean speech detected: skip ERB and DF.
-                return (false, false, false)
-            }
-            if lsnr > config.maxDbDfThresh {
-                // Mild noise: apply ERB gains only.
-                return (true, false, false)
-            }
-            // Regular noisy speech: apply both stages.
-            return (true, false, true)
-        }
-
-        private func applyConvLast(
-            input: MLXArray,
-            prefix: String,
-            main: Int,
-            pointwise: Int?,
-            bn: Int,
-            fstride: Int
-        ) throws -> MLXArray {
-            var y = input
-            y = try model.conv2dLayer(
-                y,
-                weightKey: "\(prefix).\(main).weight",
-                bias: nil,
-                fstride: fstride,
-                lookahead: 0
-            )
-            if let pointwise {
-                y = try model.conv2dLayer(
-                    y,
-                    weightKey: "\(prefix).\(pointwise).weight",
-                    bias: nil,
-                    fstride: 1,
-                    lookahead: 0
-                )
-            }
-            y = try model.batchNorm(y, prefix: "\(prefix).\(bn)")
-            y = relu(y)
-            let t = y.shape[2]
-            return y[0..., 0..., (t - 1)..<t, 0...]
-        }
-
-        private func squeezedGRUStep(
-            _ x: MLXArray,
-            gru: StreamGRU,
-            hiddenSize: Int,
-            state: inout [MLXArray]?
-        ) -> MLXArray {
-            var y = relu(model.groupedLinear(x, weight: gru.linearInWeight))
-            var nextState = [MLXArray]()
-            nextState.reserveCapacity(gru.layers.count)
-            for (layer, layerDef) in gru.layers.enumerated() {
-                let prevState: MLXArray
-                if let state, layer < state.count {
-                    prevState = state[layer]
-                } else {
-                    prevState = MLXArray.zeros([y.shape[0], hiddenSize], type: Float.self)
-                }
-                let h = gruLayerStep(y, layer: layerDef, hiddenSize: hiddenSize, prevState: prevState)
-                nextState.append(h)
-                y = h.expandedDimensions(axis: 1)
-            }
-
-            state = nextState
-            if let linearOut = gru.linearOutWeight {
-                y = relu(model.groupedLinear(y, weight: linearOut))
-            }
-            return y
-        }
-
-        private func gruLayerStep(
-            _ x: MLXArray,
-            layer: StreamGRULayer,
-            hiddenSize: Int,
-            prevState: MLXArray
-        ) -> MLXArray {
-            let xt = x[0..., 0, 0...]
-            let gx = MLX.addMM(layer.bih, xt, layer.wihT)
-            let gh = MLX.addMM(layer.bhh, prevState, layer.whhT)
-
-            let xr = gx[0..., 0..<hiddenSize]
-            let xz = gx[0..., hiddenSize..<(2 * hiddenSize)]
-            let xn = gx[0..., (2 * hiddenSize)...]
-            let hr = gh[0..., 0..<hiddenSize]
-            let hz = gh[0..., hiddenSize..<(2 * hiddenSize)]
-            let hn = gh[0..., (2 * hiddenSize)...]
-
-            let r = sigmoid(xr + hr)
-            let z = sigmoid(xz + hz)
-            let n = tanh(xn + r * hn)
-            return (MLXArray(Float(1.0)) - z) * n + z * prevState
-        }
-
-        private func erbDecoderStep(
-            emb: MLXArray,
-            e3: MLXArray,
-            e2: MLXArray,
-            e1: MLXArray,
-            e0: MLXArray
-        ) throws -> MLXArray {
-            var embDec = squeezedGRUStep(
-                emb,
-                gru: erbDecEmbGRU,
-                hiddenSize: model.config.embHiddenDim,
-                state: &erbDecState
-            )
-            let f8 = e3.shape[3]
-            embDec = embDec.reshaped([1, 1, f8, -1]).transposed(0, 3, 1, 2)
-
-            var d3 = relu(try model.applyPathwayConv(e3, prefix: "erb_dec.conv3p")) + embDec
-            d3 = relu(try model.applyRegularBlock(d3, prefix: "erb_dec.convt3"))
-            var d2 = relu(try model.applyPathwayConv(e2, prefix: "erb_dec.conv2p")) + d3
-            d2 = relu(try model.applyTransposeBlock(d2, prefix: "erb_dec.convt2", fstride: 2))
-            var d1 = relu(try model.applyPathwayConv(e1, prefix: "erb_dec.conv1p")) + d2
-            d1 = relu(try model.applyTransposeBlock(d1, prefix: "erb_dec.convt1", fstride: 2))
-            let d0 = relu(try model.applyPathwayConv(e0, prefix: "erb_dec.conv0p")) + d1
-            let out = try model.applyOutputConv(d0, prefix: "erb_dec.conv0_out")
-            return sigmoid(out)
-        }
-
-        private func dfDecoderStep(emb: MLXArray, c0: MLXArray) throws -> MLXArray {
-            var c = squeezedGRUStep(
-                emb,
-                gru: dfDecGRU,
-                hiddenSize: model.config.dfHiddenDim,
-                state: &dfDecState
-            )
-            if let dfDecSkipWeight {
-                c = c + model.groupedLinear(emb, weight: dfDecSkipWeight)
-            }
-
-            dfConvpHistory = appendHistoryFrame(dfConvpHistory, frame: c0)
-            let c0Seq = dfConvpHistory
-            var c0p = try model.conv2dLayer(
-                c0Seq,
-                weightKey: "df_dec.df_convp.1.weight",
-                bias: nil,
-                fstride: 1,
-                lookahead: 0
-            )
-            c0p = try model.conv2dLayer(
-                c0p,
-                weightKey: "df_dec.df_convp.2.weight",
-                bias: nil,
-                fstride: 1,
-                lookahead: 0
-            )
-            c0p = relu(try model.batchNorm(c0p, prefix: "df_dec.df_convp.3"))
-            let t = c0p.shape[2]
-            c0p = c0p[0..., 0..., (t - 1)..<t, 0...]
-            c0p = c0p.transposed(0, 2, 3, 1)
-
-            let dfOut = tanh(model.groupedLinear(c, weight: dfDecOutWeight))
-                .reshaped([1, 1, nbDf, dfOrder * 2])
-            return dfOut + c0p
-        }
-
-        private func applyMask(spec: MLXArray, mask: MLXArray) -> MLXArray {
-            let b = mask.shape[0]
-            let t = mask.shape[2]
-            let e = mask.shape[3]
-            let flat = mask.reshaped([b * t, e])
-            let gains = MLX.matmul(flat, model.erbInvFB).reshaped([b, 1, t, model.config.freqBins, 1])
-            return spec * gains
-        }
-
-        private func deepFilterAssign(
-            spec: MLXArray,
-            specMasked: MLXArray,
-            dfCoefs: MLXArray,
-            targetFrameIndex _: Int
-        ) throws -> MLXArray {
-            let specLow = dfSpecHistory
-            let coef = dfCoefs[0, 0..., 0, 0..<nbDf, 0...]
-
-            let sr = specLow[0..., 0..., 0]
-            let si = specLow[0..., 0..., 1]
-            let cr = coef[0..., 0..., 0]
-            let ci = coef[0..., 0..., 1]
-
-            let outReal = MLX.sum(sr * cr - si * ci, axis: 0)
-            let outImag = MLX.sum(sr * ci + si * cr, axis: 0)
-
-            let low = MLX.stacked([outReal, outImag], axis: -1)
-                .expandedDimensions(axis: 0)
-                .expandedDimensions(axis: 0)
-                .expandedDimensions(axis: 0)
-
-            if model.config.encConcat {
-                let high = specMasked[0..., 0..., 0..., nbDf..., 0...]
-                return MLX.concatenated([low, high], axis: 3)
-            }
-
-            let highUnmasked = spec[0..., 0..., 0..., nbDf..., 0...]
-            let specDf = MLX.concatenated([low, highUnmasked], axis: 3)
-            let lowAssigned = specDf[0..., 0..., 0..., 0..<nbDf, 0...]
-            let highMasked = specMasked[0..., 0..., 0..., nbDf..., 0...]
-            return MLX.concatenated([lowAssigned, highMasked], axis: 3)
-        }
-
-        private func appendHistoryFrame(_ history: MLXArray, frame: MLXArray) -> MLXArray {
-            let t = history.shape[2]
-            guard t > 1 else { return frame }
-            return MLX.concatenated([history[0..., 0..., 1..<t, 0...], frame], axis: 2)
-        }
-
-        private func updateDfSpecHistory(targetFrameIndex: Int) {
-            if !dfSpecHistoryInitialized {
-                var frames = [MLXArray]()
-                frames.reserveCapacity(dfOrder)
-                for k in 0..<dfOrder {
-                    let absoluteIndex = targetFrameIndex - dfSpecLeft + k
-                    if let frame = specRing.get(absoluteIndex: absoluteIndex) {
-                        frames.append(frame[0..<nbDf, 0...])
-                    } else {
-                        frames.append(zeroSpecLowFrame)
-                    }
-                }
-                dfSpecHistory = MLX.stacked(frames, axis: 0)
-                dfSpecHistoryInitialized = true
-                return
-            }
-            let newRightIndex = targetFrameIndex - dfSpecLeft + dfOrder - 1
-            let nextLow: MLXArray
-            if let frame = specRing.get(absoluteIndex: newRightIndex) {
-                nextLow = frame[0..<nbDf, 0...]
-            } else {
-                nextLow = zeroSpecLowFrame
-            }
-            dfSpecHistory = MLX.concatenated([dfSpecHistory[1..<dfOrder, 0..., 0...], nextLow.expandedDimensions(axis: 0)], axis: 0)
-        }
-
-        // Materialize recurrent tensors each hop so MLX does not keep growing a long lazy graph.
-        private func materializeStreamingState(output: MLXArray) {
-            eval(output, analysisMem, synthMem, erbState, dfState)
-            if let encEmbState {
-                for x in encEmbState { eval(x) }
-            }
-            if let erbDecState {
-                for x in erbDecState { eval(x) }
-            }
-            if let dfDecState {
-                for x in dfDecState { eval(x) }
-            }
-        }
-
-        private static func linspace(start: Float, end: Float, count: Int) -> [Float] {
-            guard count > 1 else { return [start] }
-            let step = (end - start) / Float(count - 1)
-            return (0..<count).map { start + Float($0) * step }
-        }
-    }
-
-    // MARK: - Network Forward
-
-    private func forward(
-        spec: MLXArray,
-        featErb: MLXArray,
-        featSpec5D: MLXArray
-    ) throws -> (MLXArray, MLXArray, MLXArray, MLXArray) {
-        let featSpec = featSpec5D
-            .squeezed(axis: 1)
-            .transposed(0, 3, 1, 2)
-
-        let featErbShift = applyLookahead(feature: featErb, lookahead: config.convLookahead)
-        let featSpecShift = applyLookahead(feature: featSpec, lookahead: config.convLookahead)
-
-        let (e0, e1, e2, e3, emb, c0, lsnr) = try encode(featErb: featErbShift, featSpec: featSpecShift)
-
-        let mask = try decodeErb(emb: emb, e3: e3, e2: e2, e1: e1, e0: e0)
-        let specMasked = applyMask(spec: spec, mask: mask)
-
-        let dfCoefs = try decodeDf(emb: emb, c0: c0)
-        let b = dfCoefs.shape[0]
-        let t = dfCoefs.shape[1]
-        let dfCoefs5 = dfCoefs
-            .reshaped([b, t, config.nbDf, config.dfOrder, 2])
-            .transposed(0, 3, 1, 2, 4)
-
-        let specEnhanced: MLXArray
-        if config.encConcat {
-            specEnhanced = deepFilter(spec: specMasked, coefs: dfCoefs5)
-        } else {
-            let specDf = deepFilter(spec: spec, coefs: dfCoefs5)
-            let low = specDf[0..., 0..., 0..., 0..<config.nbDf, 0...]
-            let high = specMasked[0..., 0..., 0..., config.nbDf..., 0...]
-            specEnhanced = MLX.concatenated([low, high], axis: 3)
-        }
-
-        return (specEnhanced, mask, lsnr, dfCoefs5)
-    }
-
-    private func forwardV1(
-        spec: MLXArray,
-        featErb: MLXArray,
-        featSpec5D: MLXArray
-    ) throws -> (MLXArray, MLXArray, MLXArray, MLXArray) {
-        let featSpec = featSpec5D
-            .squeezed(axis: 1)
-            .transposed(0, 3, 1, 2)
-
-        let (e0, e1, e2, e3, emb, c0, lsnr) = try encodeV1(featErb: featErb, featSpec: featSpec)
-        eval(emb, c0, lsnr)
-
-        var mask = try decodeErbV1(emb: emb, e3: e3, e2: e2, e1: e1, e0: e0)
-        mask = alignTimeAxis(mask, target: spec.shape[2], fillValue: 1.0, axis: 2)
-        let specMasked = applyMask(spec: spec, mask: mask)
-        eval(mask, specMasked)
-
-        var (dfCoefsBTOF2, dfAlpha) = try decodeDfV1(emb: emb, c0: c0)
-        dfCoefsBTOF2 = alignTimeAxis(dfCoefsBTOF2, target: spec.shape[2], fillValue: 0.0, axis: 1)
-        dfAlpha = alignTimeAxis(dfAlpha, target: spec.shape[2], fillValue: 0.0, axis: 1)
-        let dfCoefs5 = dfCoefsBTOF2.transposed(0, 2, 1, 3, 4)
-        eval(dfCoefs5, dfAlpha)
-
-        let specEnhanced = deepFilter(spec: specMasked, coefs: dfCoefs5, alpha: dfAlpha)
-        return (specEnhanced, mask, lsnr, dfCoefs5)
-    }
-
-    private func encodeV1(featErb: MLXArray, featSpec: MLXArray)
-        throws -> (MLXArray, MLXArray, MLXArray, MLXArray, MLXArray, MLXArray, MLXArray)
-    {
-        let e0 = try applyV1ConvKxF(
-            featErb,
-            prefix: "enc.erb_conv0",
-            fstride: 1,
-            lookahead: config.convLookahead > 0 ? 1 : 0,
-            activation: .relu
-        )
-        let e1 = try applyV1ConvKxF(
-            e0,
-            prefix: "enc.erb_conv1",
-            fstride: 2,
-            lookahead: config.convLookahead > 1 ? 1 : 0,
-            activation: .relu
-        )
-        let e2 = try applyV1ConvKxF(
-            e1,
-            prefix: "enc.erb_conv2",
-            fstride: 2,
-            lookahead: config.convLookahead > 2 ? 1 : 0,
-            activation: .relu
-        )
-        let e3 = try applyV1ConvKxF(
-            e2,
-            prefix: "enc.erb_conv3",
-            fstride: 1,
-            lookahead: 0,
-            activation: .relu
-        )
-
-        let c0 = try applyV1ConvKxF(
-            featSpec,
-            prefix: "enc.clc_conv0",
-            fstride: 1,
-            lookahead: config.convLookahead,
-            activation: .relu
-        )
-        let c1 = try applyV1ConvKxF(
-            c0,
-            prefix: "enc.clc_conv1",
-            fstride: 2,
-            lookahead: 0,
-            activation: .relu
-        )
-
-        let t = c1.shape[2]
-        let b = c1.shape[0]
-        var cemb = c1.transposed(2, 0, 1, 3).reshaped([t, b, -1])
-        cemb = relu(try groupedLinearV1(cemb, prefix: "enc.clc_fc_emb.layers"))
-
-        var emb = e3.transposed(2, 0, 1, 3).reshaped([t, b, -1])
-        emb = emb + cemb
-        emb = try groupedGRUV1(
-            emb,
-            prefix: "enc.emb_gru.grus",
-            numLayers: config.embNumLayers,
-            addOutputs: true,
-            shuffleBetweenLayers: config.groupShuffle
-        )
-        let embBT = emb.transposed(1, 0, 2)
-
-        let lsnr = sigmoid(linear(
-            embBT,
-            weight: try w("enc.lsnr_fc.0.weight"),
-            bias: try w("enc.lsnr_fc.0.bias")
-        )) * MLXArray(Float(config.lsnrMax - config.lsnrMin)) + MLXArray(Float(config.lsnrMin))
-
-        return (e0, e1, e2, e3, embBT, c0, lsnr)
-    }
-
-    private func decodeErbV1(
-        emb: MLXArray,
-        e3: MLXArray,
-        e2: MLXArray,
-        e1: MLXArray,
-        e0: MLXArray
-    ) throws -> MLXArray {
-        let b = emb.shape[0]
-        let t = emb.shape[1]
-        let f8 = e3.shape[3]
-
-        var embProj = relu(try groupedLinearV1(emb, prefix: "erb_dec.fc_emb.0.layers"))
-        embProj = embProj.reshaped([b, t, -1, f8]).transposed(0, 2, 1, 3)
-
-        let p3 = try applyV1ConvKxF(e3, prefix: "erb_dec.conv3p", fstride: 1, lookahead: 0, activation: .relu)
-        var d3 = alignAndAdd(p3, embProj)
-        d3 = try applyV1ConvKxF(d3, prefix: "erb_dec.convt3", fstride: 1, lookahead: 0, activation: .relu)
-
-        let p2 = try applyV1ConvKxF(e2, prefix: "erb_dec.conv2p", fstride: 1, lookahead: 0, activation: .relu)
-        var d2 = alignAndAdd(p2, d3)
-        d2 = try applyV1ConvKxF(d2, prefix: "erb_dec.convt2", fstride: 2, lookahead: 0, activation: .relu)
-
-        let p1 = try applyV1ConvKxF(e1, prefix: "erb_dec.conv1p", fstride: 1, lookahead: 0, activation: .relu)
-        var d1 = alignAndAdd(p1, d2)
-        d1 = try applyV1ConvKxF(d1, prefix: "erb_dec.convt1", fstride: 2, lookahead: 0, activation: .relu)
-
-        let p0 = try applyV1ConvKxF(e0, prefix: "erb_dec.conv0p", fstride: 1, lookahead: 0, activation: .relu)
-        let d0 = alignAndAdd(p0, d1)
-
-        return try applyV1ConvKxF(
-            d0,
-            prefix: "erb_dec.conv0_out",
-            fstride: 1,
-            lookahead: 0,
-            activation: .sigmoid,
-            applyBatchNorm: false,
-            allowBias: true
-        )
-    }
-
-    private func decodeDfV1(emb: MLXArray, c0: MLXArray) throws -> (MLXArray, MLXArray) {
-        let cTBI = try groupedGRUV1(
-            emb.transposed(1, 0, 2),
-            prefix: "clc_dec.clc_gru.grus",
-            numLayers: config.dfNumLayers,
-            addOutputs: true,
-            shuffleBetweenLayers: config.groupShuffle
-        )
-        let c = cTBI.transposed(1, 0, 2)
-
-        var c0p = try applyV1ConvKxF(
-            c0,
-            prefix: "clc_dec.clc_convp",
-            fstride: 1,
-            lookahead: 0,
-            activation: .relu
-        )
-        c0p = c0p.transposed(0, 2, 1, 3)  // [B, T, O*2, F]
-
-        let b = c.shape[0]
-        let t = c.shape[1]
-        let alpha = sigmoid(linear(
-            c,
-            weight: try w("clc_dec.clc_fc_a.0.weight"),
-            bias: try w("clc_dec.clc_fc_a.0.bias")
-        ))
-
-        var coefs = tanh(linear(
-            c,
-            weight: try w("clc_dec.clc_fc_out.0.weight"),
-            bias: try w("clc_dec.clc_fc_out.0.bias")
-        ))
-        coefs = coefs.reshaped([b, t, config.dfOrder * 2, config.nbDf]) + c0p
-        coefs = coefs.reshaped([b, t, config.dfOrder, 2, config.nbDf]).transposed(0, 1, 2, 4, 3)
-        return (coefs, alpha)
-    }
-
-    private func encode(featErb: MLXArray, featSpec: MLXArray)
-        throws -> (MLXArray, MLXArray, MLXArray, MLXArray, MLXArray, MLXArray, MLXArray)
-    {
-        let e0 = try applyEncoderConv(featErb, prefix: "enc.erb_conv0", main: 1, pointwise: nil, bn: 2, fstride: 1)
-        let e1 = try applyEncoderConv(e0, prefix: "enc.erb_conv1", main: 0, pointwise: 1, bn: 2, fstride: 2)
-        let e2 = try applyEncoderConv(e1, prefix: "enc.erb_conv2", main: 0, pointwise: 1, bn: 2, fstride: 2)
-        let e3 = try applyEncoderConv(e2, prefix: "enc.erb_conv3", main: 0, pointwise: 1, bn: 2, fstride: 1)
-
-        let c0 = try applyEncoderConv(featSpec, prefix: "enc.df_conv0", main: 1, pointwise: 2, bn: 3, fstride: 1)
-        let c1 = try applyEncoderConv(c0, prefix: "enc.df_conv1", main: 0, pointwise: 1, bn: 2, fstride: 2)
-
-        let b = c1.shape[0]
-        let t = c1.shape[2]
-        var cemb = c1.transposed(0, 2, 3, 1).reshaped([b, t, -1])
-        cemb = relu(groupedLinear(cemb, weight: try w("enc.df_fc_emb.0.weight")))
-
-        var emb = e3.transposed(0, 2, 3, 1).reshaped([b, t, -1])
-        emb = config.encConcat ? MLX.concatenated([emb, cemb], axis: -1) : (emb + cemb)
-
-        emb = try squeezedGRU(
-            emb,
-            prefix: "enc.emb_gru",
-            hiddenSize: config.embHiddenDim,
-            linearOut: true
-        )
-
-        let lsnr = sigmoid(linear(
-            emb,
-            weight: try w("enc.lsnr_fc.0.weight"),
-            bias: try w("enc.lsnr_fc.0.bias")
-        )) * MLXArray(Float(config.lsnrMax - config.lsnrMin)) + MLXArray(Float(config.lsnrMin))
-
-        return (e0, e1, e2, e3, emb, c0, lsnr)
-    }
-
-    private func decodeErb(
-        emb: MLXArray,
-        e3: MLXArray,
-        e2: MLXArray,
-        e1: MLXArray,
-        e0: MLXArray
-    ) throws -> MLXArray {
-        var embDec = try squeezedGRU(
-            emb,
-            prefix: "erb_dec.emb_gru",
-            hiddenSize: config.embHiddenDim,
-            linearOut: true
-        )
-
-        let b = embDec.shape[0]
-        let t = embDec.shape[1]
-        let f8 = e3.shape[3]
-        embDec = embDec.reshaped([b, t, f8, -1]).transposed(0, 3, 1, 2)
-
-        var d3 = relu(try applyPathwayConv(e3, prefix: "erb_dec.conv3p")) + embDec
-        // Matches MLX/PyTorch DF decoder: convt3 is a regular conv block, not transposed.
-        d3 = relu(try applyRegularBlock(d3, prefix: "erb_dec.convt3"))
-        var d2 = relu(try applyPathwayConv(e2, prefix: "erb_dec.conv2p")) + d3
-        d2 = relu(try applyTransposeBlock(d2, prefix: "erb_dec.convt2", fstride: 2))
-        var d1 = relu(try applyPathwayConv(e1, prefix: "erb_dec.conv1p")) + d2
-        d1 = relu(try applyTransposeBlock(d1, prefix: "erb_dec.convt1", fstride: 2))
-        let d0 = relu(try applyPathwayConv(e0, prefix: "erb_dec.conv0p")) + d1
-        let out = try applyOutputConv(d0, prefix: "erb_dec.conv0_out")
-        return sigmoid(out)
-    }
-
-    private func decodeDf(emb: MLXArray, c0: MLXArray) throws -> MLXArray {
-        var c = try squeezedGRU(
-            emb,
-            prefix: "df_dec.df_gru",
-            hiddenSize: config.dfHiddenDim,
-            linearOut: false
-        )
-
-        if weights["df_dec.df_skip.weight"] != nil {
-            c = c + groupedLinear(emb, weight: try w("df_dec.df_skip.weight"))
-        }
-
-        var c0p = try conv2dLayer(
-            c0,
-            weightKey: "df_dec.df_convp.1.weight",
-            bias: nil,
-            fstride: 1,
-            lookahead: 0
-        )
-        c0p = try conv2dLayer(
-            c0p,
-            weightKey: "df_dec.df_convp.2.weight",
-            bias: nil,
-            fstride: 1,
-            lookahead: 0
-        )
-        c0p = relu(try batchNorm(c0p, prefix: "df_dec.df_convp.3"))
-        c0p = c0p.transposed(0, 2, 3, 1)
-
-        let b = c.shape[0]
-        let t = c.shape[1]
-        let dfOut = tanh(groupedLinear(c, weight: try w("df_dec.df_out.0.weight")))
-            .reshaped([b, t, config.nbDf, config.dfOrder * 2])
-
-        return dfOut + c0p
-    }
-
-    private func applyMask(spec: MLXArray, mask: MLXArray) -> MLXArray {
-        let b = mask.shape[0]
-        let t = mask.shape[2]
-        let e = mask.shape[3]
-        let flat = mask.reshaped([b * t, e])
-        let gains = MLX.matmul(flat, erbInvFB).reshaped([b, 1, t, config.freqBins, 1])
-        return spec * gains
-    }
-
-    private func deepFilter(spec: MLXArray, coefs: MLXArray, alpha: MLXArray? = nil) -> MLXArray {
-        let t = spec.shape[2]
-        let padLeft = config.dfOrder - 1 - config.dfLookahead
-        let padRight = config.dfLookahead
-
-        let specLow = spec[0..., 0, 0..., 0..<config.nbDf, 0...]
-        let padded = MLX.padded(
-            specLow,
-            widths: [
-                .init(0),
-                .init((padLeft, padRight)),
-                .init(0),
-                .init(0),
-            ],
-            mode: .constant
-        )
-
-        let b = spec.shape[0]
-        var outR = MLXArray.zeros([b, t, config.nbDf], dtype: spec.dtype)
-        var outI = MLXArray.zeros([b, t, config.nbDf], dtype: spec.dtype)
-        for k in 0..<config.dfOrder {
-            let window = padded[0..., k..<(k + t), 0..., 0...]  // [B, T, F_df, 2]
-            let coef = coefs[0..., k, 0..., 0..., 0...]  // [B, T, F_df, 2]
-            let sr = window[0..., 0..., 0..., 0]
-            let si = window[0..., 0..., 0..., 1]
-            let cr = coef[0..., 0..., 0..., 0]
-            let ci = coef[0..., 0..., 0..., 1]
-
-            outR = outR + (sr * cr - si * ci)
-            outI = outI + (sr * ci + si * cr)
-        }
-
-        var low = MLX.stacked([outR, outI], axis: -1).expandedDimensions(axis: 1)
-        if let alpha {
-            let b = spec.shape[0]
-            let a = alpha.reshaped([b, 1, t, 1, 1])
-            let origLow = spec[0..., 0..., 0..., 0..<config.nbDf, 0...]
-            low = low * a + origLow * (MLXArray(Float(1.0)) - a)
-        }
-        let high = spec[0..., 0..., 0..., config.nbDf..., 0...]
-        return MLX.concatenated([low, high], axis: 3)
-    }
-
-    private enum V1Activation {
-        case none
-        case relu
-        case sigmoid
-    }
-
-    private func applyV1ConvKxF(
-        _ x: MLXArray,
-        prefix: String,
-        fstride: Int,
-        lookahead: Int,
-        activation: V1Activation,
-        applyBatchNorm: Bool = true,
-        allowBias: Bool = false
-    ) throws -> MLXArray {
-        let mainKey: String
-        let transposed: Bool
-        if weights["\(prefix).sconvt.weight"] != nil {
-            mainKey = "\(prefix).sconvt.weight"
-            transposed = true
-        } else {
-            mainKey = "\(prefix).sconv.weight"
-            transposed = false
-        }
-
-        let mainWeight = try w(mainKey)
-        var y: MLXArray
-        if transposed {
-            let groups = max(1, mainWeight.shape[0] / max(1, mainWeight.shape[1]))
-            y = try convTranspose2dLayer(
-                x,
-                weight: mainWeight,
-                fstride: fstride,
-                groups: groups
-            )
-        } else {
-            let bias = allowBias ? weights["\(prefix).sconv.bias"] : nil
-            y = try conv2dLayer(
-                x,
-                weight: mainWeight,
-                bias: bias,
-                fstride: fstride,
-                lookahead: lookahead
-            )
-        }
-
-        if weights["\(prefix).1x1conv.weight"] != nil {
-            y = try conv2dLayer(
-                y,
-                weightKey: "\(prefix).1x1conv.weight",
-                bias: nil,
-                fstride: 1,
-                lookahead: 0
-            )
-        }
-
-        if applyBatchNorm, weights["\(prefix).norm.running_mean"] != nil {
-            y = try batchNorm(y, prefix: "\(prefix).norm")
-        }
-
-        switch activation {
-        case .none:
-            return y
-        case .relu:
-            return relu(y)
-        case .sigmoid:
-            return sigmoid(y)
-        }
-    }
-
-    private func groupedLinearV1(
-        _ x: MLXArray,
-        prefix: String
-    ) throws -> MLXArray {
-        if let pack = v1GroupedLinearPacks[prefix] {
-            let b = x.shape[0]
-            let t = x.shape[1]
-            var y = MLX.einsum(
-                "btgi,gio->btgo",
-                x.reshaped([b, t, pack.groups, pack.inputPerGroup]),
-                pack.weightGIO
-            ) + pack.biasGO.reshaped([1, 1, pack.groups, pack.outputPerGroup])
-            y = y.reshaped([b, t, pack.groups * pack.outputPerGroup])
-            if config.groupShuffle, pack.groups > 1 {
-                let hiddenPerGroup = y.shape[2] / pack.groups
-                y = y
-                    .reshaped([y.shape[0], y.shape[1], hiddenPerGroup, pack.groups])
-                    .transposed(0, 1, 3, 2)
-                    .reshaped([y.shape[0], y.shape[1], -1])
-            }
-            return y
-        }
-
-        let groups = max(1, config.linearGroups)
-        var ys = [MLXArray]()
-        ys.reserveCapacity(groups)
-        for g in 0..<groups {
-            let wg = try w("\(prefix).\(g).weight")
-            let bg = try w("\(prefix).\(g).bias")
-            let inPerGroup = wg.shape[1]
-            let start = g * inPerGroup
-            let stop = start + inPerGroup
-            let xg = x[0..., 0..., start..<stop]
-            let x2 = xg.reshaped([-1, inPerGroup])
-            let y2 = MLX.addMM(bg, x2, wg.transposed())
-            ys.append(y2.reshaped([x.shape[0], x.shape[1], wg.shape[0]]))
-        }
-        var y = MLX.concatenated(ys, axis: 2)
-        if config.groupShuffle, groups > 1 {
-            let hiddenPerGroup = y.shape[2] / groups
-            y = y
-                .reshaped([y.shape[0], y.shape[1], hiddenPerGroup, groups])
-                .transposed(0, 1, 3, 2)
-                .reshaped([y.shape[0], y.shape[1], -1])
-        }
-        return y
-    }
-
-    private func groupedGRUV1(
-        _ xTBI: MLXArray,
-        prefix: String,
-        numLayers: Int,
-        addOutputs: Bool,
-        shuffleBetweenLayers: Bool
-    ) throws -> MLXArray {
-        if let pack = v1GroupedGRUPacks[prefix], pack.layers.count >= numLayers {
-            return groupedGRUV1Packed(
-                xTBI,
-                pack: pack,
-                numLayers: numLayers,
-                addOutputs: addOutputs,
-                shuffleBetweenLayers: shuffleBetweenLayers
-            )
-        }
-
-        let groups = max(1, config.gruGroups)
-        var cur = xTBI
-        var out = xTBI
-
-        for layer in 0..<numLayers {
-            let base = "\(prefix).\(layer).layers"
-            let w0 = try w("\(base).0.weight_ih_l0")
-            let inPerGroup = w0.shape[1]
-            let hiddenPerGroup = w0.shape[0] / 3
-            var ys = [MLXArray]()
-            ys.reserveCapacity(groups)
-            for g in 0..<groups {
-                let start = g * inPerGroup
-                let stop = start + inPerGroup
-                let xg = cur[0..., 0..., start..<stop]
-                let yg = try gruCellSequenceV1(
-                    xg,
-                    weightIH: w("\(base).\(g).weight_ih_l0"),
-                    weightHH: w("\(base).\(g).weight_hh_l0"),
-                    biasIH: w("\(base).\(g).bias_ih_l0"),
-                    biasHH: w("\(base).\(g).bias_hh_l0"),
-                    hiddenSize: hiddenPerGroup
-                )
-                ys.append(yg)
-            }
-
-            var layerOut = MLX.concatenated(ys, axis: 2)
-            if shuffleBetweenLayers && layer < numLayers - 1 && groups > 1 {
-                let hidden = layerOut.shape[2] / groups
-                layerOut = layerOut
-                    .reshaped([layerOut.shape[0], layerOut.shape[1], hidden, groups])
-                    .transposed(0, 1, 3, 2)
-                    .reshaped([layerOut.shape[0], layerOut.shape[1], -1])
-            }
-
-            if addOutputs {
-                out = (layer == 0) ? layerOut : (out + layerOut)
-            } else {
-                out = layerOut
-            }
-            cur = layerOut
-        }
-
-        return out
-    }
-
-    private func groupedGRUV1Packed(
-        _ xTBI: MLXArray,
-        pack: V1GroupedGRUPack,
-        numLayers: Int,
-        addOutputs: Bool,
-        shuffleBetweenLayers: Bool
-    ) -> MLXArray {
-        let groups = pack.groups
-        let t = xTBI.shape[0]
-        let b = xTBI.shape[1]
-
-        var curMLX = xTBI
-        var outFlat: [Float]? = nil
-
-        for layer in 0..<numLayers {
-            let p = pack.layers[layer]
-            let h = p.hiddenPerGroup
-            let inpg = p.inputPerGroup
-            let h3 = 3 * h
-            let totalH = groups * h
-
-            // Batch input projection on GPU: [T, B, G, I] @ [G, I, 3H] -> [T, B, G, 3H]
-            let x4 = curMLX.reshaped([t, b, groups, inpg])
-            let gxAllMLX = MLX.einsum("tbgi,gio->tbgo", x4, p.weightIHGI3H)
-                + p.biasIHG3H.reshaped([1, 1, groups, h3])
-            eval(gxAllMLX)
-
-            // Extract to CPU: gxAll is [T, B, G, 3H] contiguous
-            let gxAll = gxAllMLX.asType(.float32).reshaped([-1]).asArray(Float.self)
-
-            // Extract hidden weights to CPU: wHH [G, H, 3H], bHH [G, 3H]
-            let wHH = p.weightHHGH3H.asType(.float32).reshaped([-1]).asArray(Float.self)
-            let bHH = p.biasHHG3H.asType(.float32).reshaped([-1]).asArray(Float.self)
-
-            var output = Array<Float>(repeating: 0, count: t * b * totalH)
-            var state = Array<Float>(repeating: 0, count: b * totalH)
-            var gh = Array<Float>(repeating: 0, count: h3)
-
-            for ti in 0..<t {
-                for bi in 0..<b {
-                    for g in 0..<groups {
-                        let stOff = bi * totalH + g * h
-                        let gxOff = ((ti * b + bi) * groups + g) * h3
-                        let wHHOff = g * h * h3
-                        let bHHOff = g * h3
-
-                        // gh = state[stOff:stOff+h] @ wHH[g]
-                        // wHH[g] is [H, 3H] row-major; we want gh[o] = sum_h state[h] * wHH[h,o]
-                        // C[1,3H] = A[1,H] × B[H,3H]
-                        state.withUnsafeBufferPointer { sBuf in
-                            wHH.withUnsafeBufferPointer { wBuf in
-                                gh.withUnsafeMutableBufferPointer { gBuf in
-                                    vDSP_mmul(
-                                        sBuf.baseAddress! + stOff, 1,
-                                        wBuf.baseAddress! + wHHOff, 1,
-                                        gBuf.baseAddress!, 1,
-                                        vDSP_Length(1), vDSP_Length(h3), vDSP_Length(h)
-                                    )
-                                }
-                            }
-                        }
-
-                        // GRU gates
-                        for k in 0..<h {
-                            let xr = gxAll[gxOff + k]
-                            let xz = gxAll[gxOff + h + k]
-                            let xn = gxAll[gxOff + 2 * h + k]
-                            let hr = gh[k] + bHH[bHHOff + k]
-                            let hz = gh[h + k] + bHH[bHHOff + h + k]
-                            let hn = gh[2 * h + k] + bHH[bHHOff + 2 * h + k]
-
-                            let r = 1.0 / (1.0 + expf(-(xr + hr)))
-                            let z = 1.0 / (1.0 + expf(-(xz + hz)))
-                            let n = tanhf(xn + r * hn)
-                            let prev = state[stOff + k]
-                            state[stOff + k] = n + z * (prev - n)
-                        }
-
-                        // Copy state to output
-                        let outOff = (ti * b + bi) * totalH + g * h
-                        output.replaceSubrange(outOff..<(outOff + h), with: state[stOff..<(stOff + h)])
-                    }
-                }
-
-            }
-
-            // Group shuffle between layers (not last layer)
-            // Matches: reshape([H, G]).transpose(G, H).reshape(-1)
-            // Mapping: source flat index i → dest ((i % G) * H + (i / G))
-            if shuffleBetweenLayers && layer < numLayers - 1 && groups > 1 {
-                var shuffled = Array<Float>(repeating: 0, count: output.count)
-                for ti in 0..<t {
-                    for bi in 0..<b {
-                        let base = (ti * b + bi) * totalH
-                        for i in 0..<totalH {
-                            let dest = (i % groups) * h + (i / groups)
-                            shuffled[base + dest] = output[base + i]
-                        }
-                    }
-                }
-                output = shuffled
-            }
-
-            // Accumulate outputs
-            if addOutputs {
-                if layer == 0 {
-                    outFlat = output
-                } else {
-                    for i in 0..<output.count {
-                        outFlat![i] += output[i]
-                    }
-                }
-            } else {
-                outFlat = output
-            }
-
-            // For next layer, convert back to MLXArray
-            if layer < numLayers - 1 {
-                curMLX = MLXArray(output).reshaped([t, b, totalH])
-            }
-        }
-
-        return MLXArray(outFlat!).reshaped([t, b, groups * pack.layers[0].hiddenPerGroup])
-    }
-
-    private func gruCellSequenceV1(
-        _ xTBI: MLXArray,
-        weightIH: @autoclosure () throws -> MLXArray,
-        weightHH: @autoclosure () throws -> MLXArray,
-        biasIH: @autoclosure () throws -> MLXArray,
-        biasHH: @autoclosure () throws -> MLXArray,
-        hiddenSize: Int
-    ) throws -> MLXArray {
-        let wihT = try weightIH().transposed()
-        let whhT = try weightHH().transposed()
-        let bih = try biasIH()
-        let bhh = try biasHH()
-
-        var h = MLXArray.zeros([xTBI.shape[1], hiddenSize], type: Float.self)
-        var states = [MLXArray]()
-        states.reserveCapacity(xTBI.shape[0])
-
-        for i in 0..<xTBI.shape[0] {
-            let xt = xTBI[i, 0..., 0...]
-            let gx = MLX.addMM(bih, xt, wihT)
-            let gh = MLX.addMM(bhh, h, whhT)
-
-            let xr = gx[0..., 0..<hiddenSize]
-            let xz = gx[0..., hiddenSize..<(2 * hiddenSize)]
-            let xn = gx[0..., (2 * hiddenSize)...]
-            let hr = gh[0..., 0..<hiddenSize]
-            let hz = gh[0..., hiddenSize..<(2 * hiddenSize)]
-            let hn = gh[0..., (2 * hiddenSize)...]
-
-            let r = sigmoid(xr + hr)
-            let z = sigmoid(xz + hz)
-            let n = tanh(xn + r * hn)
-            h = n + z * (h - n)
-            states.append(h)
-        }
-        return MLX.stacked(states, axis: 0)
-    }
-
-    private func alignAndAdd(_ a: MLXArray, _ b: MLXArray) -> MLXArray {
-        let t = min(a.shape[2], b.shape[2])
-        let f = min(a.shape[3], b.shape[3])
-        return a[0..., 0..., 0..<t, 0..<f] + b[0..., 0..., 0..<t, 0..<f]
-    }
-
-    private func alignTimeAxis(
-        _ x: MLXArray,
-        target: Int,
-        fillValue: Float,
-        axis: Int
-    ) -> MLXArray {
-        let t = x.shape[axis]
-        if t == target { return x }
-        if t > target {
-            switch axis {
-            case 1:
-                if x.ndim == 3 {
-                    return x[0..., 0..<target, 0...]
-                }
-                return x[0..., 0..<target, 0..., 0..., 0...]
-            case 2:
-                if x.ndim == 4 {
-                    return x[0..., 0..., 0..<target, 0...]
-                }
-                return x[0..., 0..., 0..<target, 0..., 0...]
-            default:
-                return x
-            }
-        }
-        switch axis {
-        case 1:
-            if x.ndim == 3 {
-                let pad = MLX.full([x.shape[0], target - t, x.shape[2]], values: fillValue)
-                return MLX.concatenated([x, pad], axis: 1)
-            }
-            let pad = MLX.full([x.shape[0], target - t, x.shape[2], x.shape[3], x.shape[4]], values: fillValue)
-            return MLX.concatenated([x, pad], axis: 1)
-        case 2:
-            if x.ndim == 4 {
-                let pad = MLX.full([x.shape[0], x.shape[1], target - t, x.shape[3]], values: fillValue)
-                return MLX.concatenated([x, pad], axis: 2)
-            }
-            let pad = MLX.full([x.shape[0], x.shape[1], target - t, x.shape[3], x.shape[4]], values: fillValue)
-            return MLX.concatenated([x, pad], axis: 2)
-        default:
-            return x
-        }
-    }
-
-    // MARK: - Layer Helpers
-
-    private func applyEncoderConv(
-        _ x: MLXArray,
-        prefix: String,
-        main: Int,
-        pointwise: Int?,
-        bn: Int,
-        fstride: Int
-    ) throws -> MLXArray {
-        var y = try conv2dLayer(
-            x,
-            weightKey: "\(prefix).\(main).weight",
-            bias: nil,
-            fstride: fstride,
-            lookahead: 0
-        )
-        if let pointwise {
-            y = try conv2dLayer(
-                y,
-                weightKey: "\(prefix).\(pointwise).weight",
-                bias: nil,
-                fstride: 1,
-                lookahead: 0
-            )
-        }
-        y = try batchNorm(y, prefix: "\(prefix).\(bn)")
-        return relu(y)
-    }
-
-    private func applyPathwayConv(_ x: MLXArray, prefix: String) throws -> MLXArray {
-        var y = try conv2dLayer(
-            x,
-            weightKey: "\(prefix).0.weight",
-            bias: nil,
-            fstride: 1,
-            lookahead: 0
-        )
-        y = try batchNorm(y, prefix: "\(prefix).1")
-        return relu(y)
-    }
-
-    private func applyTransposeBlock(_ x: MLXArray, prefix: String, fstride: Int) throws -> MLXArray {
-        var y = try convTranspose2dLayer(
-            x,
-            weightKey: "\(prefix).0.weight",
-            fstride: fstride,
-            groups: config.convCh
-        )
-        y = try conv2dLayer(
-            y,
-            weightKey: "\(prefix).1.weight",
-            bias: nil,
-            fstride: 1,
-            lookahead: 0
-        )
-        return try batchNorm(y, prefix: "\(prefix).2")
-    }
-
-    private func applyRegularBlock(_ x: MLXArray, prefix: String) throws -> MLXArray {
-        var y = try conv2dLayer(
-            x,
-            weightKey: "\(prefix).0.weight",
-            bias: nil,
-            fstride: 1,
-            lookahead: 0
-        )
-        y = try conv2dLayer(
-            y,
-            weightKey: "\(prefix).1.weight",
-            bias: nil,
-            fstride: 1,
-            lookahead: 0
-        )
-        return try batchNorm(y, prefix: "\(prefix).2")
-    }
-
-    private func applyOutputConv(_ x: MLXArray, prefix: String) throws -> MLXArray {
-        var y = try conv2dLayer(
-            x,
-            weightKey: "\(prefix).0.weight",
-            bias: nil,
-            fstride: 1,
-            lookahead: 0
-        )
-        y = try batchNorm(y, prefix: "\(prefix).1")
-        return y
-    }
-
-    private func conv2dLayer(
-        _ xBCHW: MLXArray,
-        weightKey: String,
-        bias: MLXArray?,
-        fstride: Int,
-        lookahead: Int
-    ) throws -> MLXArray {
-        if let wOHWI = conv2dWeightsOHWI[weightKey] {
-            return conv2dLayer(
-                xBCHW,
-                weightOHWI: wOHWI,
-                bias: bias,
-                fstride: fstride,
-                lookahead: lookahead
-            )
-        }
-        return try conv2dLayer(
-            xBCHW,
-            weight: w(weightKey),
-            bias: bias,
-            fstride: fstride,
-            lookahead: lookahead
-        )
-    }
-
-    private func conv2dLayer(
-        _ xBCHW: MLXArray,
-        weight: MLXArray,
-        bias: MLXArray?,
-        fstride: Int,
-        lookahead: Int
-    ) throws -> MLXArray {
-        conv2dLayer(
-            xBCHW,
-            weightOHWI: weight.transposed(0, 2, 3, 1),
-            bias: bias,
-            fstride: fstride,
-            lookahead: lookahead
-        )
-    }
-
-    private func conv2dLayer(
-        _ xBCHW: MLXArray,
-        weightOHWI: MLXArray,
-        bias: MLXArray?,
-        fstride: Int,
-        lookahead: Int
-    ) -> MLXArray {
-        let kT = weightOHWI.shape[1]
-        let kF = weightOHWI.shape[2]
-        let inPerGroup = weightOHWI.shape[3]
-        let inChannels = xBCHW.shape[1]
-        let groups = max(1, inChannels / max(1, inPerGroup))
-
-        let rawLeft = kT - 1 - lookahead
-        let timeCrop = max(0, -rawLeft)
-        let timePadLeft = max(0, rawLeft)
-        let timePadRight = max(0, lookahead)
-        let freqPad = kF / 2
-
-        var x = xBCHW.transposed(0, 2, 3, 1)  // NHWC
-        if timeCrop > 0, x.shape[1] > timeCrop {
-            x = x[0..., timeCrop..., 0..., 0...]
-        }
-        x = MLX.padded(
-            x,
-            widths: [
-                .init(0),
-                .init((timePadLeft, timePadRight)),
-                .init((freqPad, freqPad)),
-                .init(0),
-            ],
-            mode: .constant
-        )
-
-        var y = MLX.conv2d(x, weightOHWI, stride: [1, fstride], padding: [0, 0], groups: groups)
-        if let bias {
-            y = y + bias.reshaped([1, 1, 1, bias.shape[0]])
-        }
-        return y.transposed(0, 3, 1, 2)
-    }
-
-    private func convTranspose2dLayer(
-        _ xBCHW: MLXArray,
-        weightKey: String,
-        fstride: Int,
-        groups: Int
-    ) throws -> MLXArray {
-        if groups > 1, let denseWeight = convTransposeDenseWeights[weightKey] {
-            var x = xBCHW.transposed(0, 2, 3, 1)  // NHWC
-            let kT = denseWeight.shape[1]
-            let kF = denseWeight.shape[2]
-            let padding = IntOrPair((kT - 1, kF / 2))
-            let outputPadding = IntOrPair((0, kF / 2))
-            x = MLX.convTransposed2d(
-                x,
-                denseWeight,
-                stride: [1, fstride],
-                padding: padding,
-                outputPadding: outputPadding,
-                groups: 1
-            )
-            return x.transposed(0, 3, 1, 2)
-        }
-        if groups > 1, let groupedWeights = convTransposeGroupWeights[weightKey], groupedWeights.count == groups {
-            return convTranspose2dLayer(
-                xBCHW,
-                groupedWeights: groupedWeights,
-                fstride: fstride
-            )
-        }
-        return try convTranspose2dLayer(
-            xBCHW,
-            weight: w(weightKey),
-            fstride: fstride,
-            groups: groups
-        )
-    }
-
-    private func convTranspose2dLayer(
-        _ xBCHW: MLXArray,
-        weight: MLXArray,
-        fstride: Int,
-        groups: Int
-    ) throws -> MLXArray {
-        var x = xBCHW.transposed(0, 2, 3, 1)  // NHWC
-        let kT = weight.shape[2]
-        let kF = weight.shape[3]
-        let padding = IntOrPair((kT - 1, kF / 2))
-        let outputPadding = IntOrPair((0, kF / 2))
-
-        if groups <= 1 {
-            let w = weight.transposed(1, 2, 3, 0)
-            x = MLX.convTransposed2d(
-                x,
-                w,
-                stride: [1, fstride],
-                padding: padding,
-                outputPadding: outputPadding,
-                groups: 1
-            )
-            return x.transposed(0, 3, 1, 2)
-        }
-
-        let inPerGroup = max(1, x.shape[3] / groups)
-        let outPerGroup = weight.shape[1]
-        var ys = [MLXArray]()
-        ys.reserveCapacity(groups)
-
-        for g in 0..<groups {
-            let inStart = g * inPerGroup
-            let inEnd = inStart + inPerGroup
-            let xg = x[0..., 0..., 0..., inStart..<inEnd]
-
-            let wg = weight[inStart..<inEnd, 0..., 0..., 0...]
-            let wT = wg.transposed(1, 2, 3, 0)  // [out_pg, kT, kF, in_pg]
-            let yg = MLX.convTransposed2d(
-                xg,
-                wT,
-                stride: [1, fstride],
-                padding: padding,
-                outputPadding: outputPadding,
-                groups: 1
-            )
-            ys.append(yg)
-        }
-
-        let y = MLX.concatenated(ys, axis: 3)
-        _ = outPerGroup  // keep shape intent explicit
-        return y.transposed(0, 3, 1, 2)
-    }
-
-    private func convTranspose2dLayer(
-        _ xBCHW: MLXArray,
-        groupedWeights: [MLXArray],
-        fstride: Int
-    ) -> MLXArray {
-        let groups = groupedWeights.count
-        var x = xBCHW.transposed(0, 2, 3, 1)  // NHWC
-        let kT = groupedWeights[0].shape[1]
-        let kF = groupedWeights[0].shape[2]
-        let padding = IntOrPair((kT - 1, kF / 2))
-        let outputPadding = IntOrPair((0, kF / 2))
-        let inPerGroup = max(1, x.shape[3] / groups)
-
-        var ys = [MLXArray]()
-        ys.reserveCapacity(groups)
-        for (g, wT) in groupedWeights.enumerated() {
-            let inStart = g * inPerGroup
-            let inEnd = inStart + inPerGroup
-            let xg = x[0..., 0..., 0..., inStart..<inEnd]
-            let yg = MLX.convTransposed2d(
-                xg,
-                wT,
-                stride: [1, fstride],
-                padding: padding,
-                outputPadding: outputPadding,
-                groups: 1
-            )
-            ys.append(yg)
-        }
-        x = MLX.concatenated(ys, axis: 3)
-        return x.transposed(0, 3, 1, 2)
-    }
-
-    private func groupedConv2d(
-        input: MLXArray,
-        weight: MLXArray,
-        strideW: Int,
-        groups: Int
-    ) throws -> MLXArray {
-        MLX.conv2d(input, weight, stride: [1, strideW], padding: [0, 0], groups: groups)
-    }
-
-    private func batchNorm(_ x: MLXArray, prefix: String) throws -> MLXArray {
-        if let scale = bnScale[prefix], let bias = bnBias[prefix] {
-            return x * scale + bias
-        }
-
-        let gamma = try w("\(prefix).weight")
-        let beta = try w("\(prefix).bias")
-        let mean = try w("\(prefix).running_mean")
-        let variance = try w("\(prefix).running_var")
-        let scale = (gamma / MLX.sqrt(variance + MLXArray(Float(1e-5))))
-            .reshaped([1, gamma.shape[0], 1, 1])
-        let shift = (beta - mean * (gamma / MLX.sqrt(variance + MLXArray(Float(1e-5)))))
-            .reshaped([1, beta.shape[0], 1, 1])
-        return x * scale + shift
-    }
-
-    private func squeezedGRU(
-        _ x: MLXArray,
-        prefix: String,
-        hiddenSize: Int,
-        linearOut: Bool
-    ) throws -> MLXArray {
-        var y = relu(groupedLinear(x, weight: try w("\(prefix).linear_in.0.weight")))
-
-        var layer = 0
-        while weights["\(prefix).gru.weight_ih_l\(layer)"] != nil {
-            y = try pytorchGRULayer(y, prefix: "\(prefix).gru", layer: layer, hiddenSize: hiddenSize)
-            layer += 1
-        }
-
-        if linearOut, weights["\(prefix).linear_out.0.weight"] != nil {
-            y = relu(groupedLinear(y, weight: try w("\(prefix).linear_out.0.weight")))
-        }
-        return y
-    }
-
-    private func groupedLinear(_ x: MLXArray, weight: MLXArray) -> MLXArray {
-        let groups = weight.shape[0]
-        let ws = weight.shape[1]
-        let hs = weight.shape[2]
-        let b = x.shape[0]
-        let t = x.shape[1]
-        let reshaped = x.reshaped([b, t, groups, ws])
-        let out = MLX.einsum("btgi,gih->btgh", reshaped, weight)
-        return out.reshaped([b, t, groups * hs])
-    }
-
-    private func pytorchGRULayer(
-        _ x: MLXArray,
-        prefix: String,
-        layer: Int,
-        hiddenSize: Int
-    ) throws -> MLXArray {
-        let wihKey = "\(prefix).weight_ih_l\(layer)"
-        let whhKey = "\(prefix).weight_hh_l\(layer)"
-        let wihT: MLXArray
-        if let cached = gruTransposedWeights[wihKey] {
-            wihT = cached
-        } else {
-            wihT = try w(wihKey).transposed()
-        }
-        let whhT: MLXArray
-        if let cached = gruTransposedWeights[whhKey] {
-            whhT = cached
-        } else {
-            whhT = try w(whhKey).transposed()
-        }
-        let bih = try w("\(prefix).bias_ih_l\(layer)")
-        let bhh = try w("\(prefix).bias_hh_l\(layer)")
-
-        let batchSize = x.shape[0]
-        let t = x.shape[1]
-        let h3 = 3 * hiddenSize
-
-        // Batch input projection on GPU: gxAll[B, T, 3H] = x @ wihT + bih
-        let x2 = x.reshaped([batchSize * t, x.shape[2]])
-        let gxAllMLX = (MLX.matmul(x2, wihT) + bih).reshaped([batchSize, t, h3])
-        eval(gxAllMLX)
-
-        // Extract to CPU
-        let gxAll = gxAllMLX.asType(.float32).reshaped([-1]).asArray(Float.self)
-        let wHH = whhT.asType(.float32).reshaped([-1]).asArray(Float.self)
-        let biasHH = bhh.asType(.float32).reshaped([-1]).asArray(Float.self)
-
-        let totalOut = batchSize * t * hiddenSize
-        var output = Array<Float>(repeating: 0, count: totalOut)
-        var state = Array<Float>(repeating: 0, count: batchSize * hiddenSize)
-        var ghBuf = Array<Float>(repeating: 0, count: h3)
-
-        for ti in 0..<t {
-            for bi in 0..<batchSize {
-                let stOff = bi * hiddenSize
-                let gxOff = (bi * t + ti) * h3
-
-                // gh = state @ wHH (hidden projection via Accelerate)
-                state.withUnsafeBufferPointer { sBuf in
-                    wHH.withUnsafeBufferPointer { wBuf in
-                        ghBuf.withUnsafeMutableBufferPointer { gBuf in
-                            vDSP_mmul(
-                                sBuf.baseAddress! + stOff, 1,
-                                wBuf.baseAddress!, 1,
-                                gBuf.baseAddress!, 1,
-                                vDSP_Length(1), vDSP_Length(h3), vDSP_Length(hiddenSize)
-                            )
-                        }
-                    }
-                }
-
-                // GRU gates (PyTorch convention: (1-z)*n + z*h)
-                for k in 0..<hiddenSize {
-                    let xr = gxAll[gxOff + k]
-                    let xz = gxAll[gxOff + hiddenSize + k]
-                    let xn = gxAll[gxOff + 2 * hiddenSize + k]
-                    let hr = ghBuf[k] + biasHH[k]
-                    let hz = ghBuf[hiddenSize + k] + biasHH[hiddenSize + k]
-                    let hn = ghBuf[2 * hiddenSize + k] + biasHH[2 * hiddenSize + k]
-
-                    let r = 1.0 / (1.0 + expf(-(xr + hr)))
-                    let z = 1.0 / (1.0 + expf(-(xz + hz)))
-                    let n = tanhf(xn + r * hn)
-                    state[stOff + k] = (1.0 - z) * n + z * state[stOff + k]
-                }
-
-                let outOff = (bi * t + ti) * hiddenSize
-                output.replaceSubrange(outOff..<(outOff + hiddenSize), with: state[stOff..<(stOff + hiddenSize)])
-            }
-        }
-
-        return MLXArray(output).reshaped([batchSize, t, hiddenSize])
-    }
-
-    private func linear(_ x: MLXArray, weight: MLXArray, bias: MLXArray) -> MLXArray {
-        let b = x.shape[0]
-        let t = x.shape[1]
-        let x2 = x.reshaped([b * t, x.shape[2]])
-        var y = MLX.matmul(x2, weight.transposed())
-        y = y + bias
-        return y.reshaped([b, t, weight.shape[0]])
-    }
-
-    // MARK: - Feature Helpers
-
-    private func bandMeanNorm(_ x: MLXArray) -> MLXArray {
-        let frames = x.shape[0]
-        let a = normAlpha()
-        let oneMinusA = Float(1.0) - a
-        let time = MLXArray.arange(frames).asType(.float32)
-        let powers = MLX.pow(MLXArray(a), time) // [T]
-        let invPowers = MLXArray(Float(1.0)) / powers
-
-        let scaled = x * invPowers.expandedDimensions(axis: 1)
-        let accum = cumsum(scaled, axis: 0)
-
-        let initState = MLXArray(Self.linspace(start: -60.0, end: -90.0, count: x.shape[1]))
-            .expandedDimensions(axis: 0)
-        let state = powers.expandedDimensions(axis: 1) * (initState + MLXArray(oneMinusA) * accum)
-        return (x - state) / MLXArray(Float(40.0))
-    }
-
-    private func bandUnitNorm(real: MLXArray, imag: MLXArray) -> (MLXArray, MLXArray) {
-        let frames = real.shape[0]
-        let a = normAlpha()
-        let oneMinusA = Float(1.0) - a
-        let time = MLXArray.arange(frames).asType(.float32)
-        let powers = MLX.pow(MLXArray(a), time) // [T]
-        let invPowers = MLXArray(Float(1.0)) / powers
-
-        let mag = MLX.sqrt(real.square() + imag.square())
-        let scaled = mag * invPowers.expandedDimensions(axis: 1)
-        let accum = cumsum(scaled, axis: 0)
-
-        let initState = MLXArray(Self.linspace(start: 0.001, end: 0.0001, count: real.shape[1]))
-            .expandedDimensions(axis: 0)
-        let state = powers.expandedDimensions(axis: 1) * (initState + MLXArray(oneMinusA) * accum)
-        let denom = MLX.sqrt(MLX.maximum(state, MLXArray(Float(1e-12))))
-        return (real / denom, imag / denom)
-    }
-
-    // Exact sequential EMA path (libDF-style), primarily for DF1 parity.
-    private func bandMeanNormExact(_ x: MLXArray) -> MLXArray {
-        let frames = x.shape[0]
-        let bands = x.shape[1]
-        let a = normAlpha()
-        let oneMinusA = Float(1.0) - a
-
-        let xVals = x.asArray(Float.self)
-        var out = Array<Float>(repeating: 0, count: xVals.count)
-        var state = Self.linspace(start: -60.0, end: -90.0, count: bands)
-
-        for t in 0..<frames {
-            let base = t * bands
-            for e in 0..<bands {
-                let idx = base + e
-                let xv = xVals[idx]
-                state[e] = xv * oneMinusA + state[e] * a
-                out[idx] = (xv - state[e]) / 40.0
-            }
-        }
-        return MLXArray(out).reshaped([frames, bands])
-    }
-
-    // Exact sequential complex unit-norm path (libDF-style), primarily for DF1 parity.
-    private func bandUnitNormExact(real: MLXArray, imag: MLXArray) -> (MLXArray, MLXArray) {
-        let frames = real.shape[0]
-        let freqs = real.shape[1]
-        let a = normAlpha()
-        let oneMinusA = Float(1.0) - a
-
-        let rVals = real.asArray(Float.self)
-        let iVals = imag.asArray(Float.self)
-        var outR = Array<Float>(repeating: 0, count: rVals.count)
-        var outI = Array<Float>(repeating: 0, count: iVals.count)
-        var state = Self.linspace(start: 0.001, end: 0.0001, count: freqs)
-
-        for t in 0..<frames {
-            let base = t * freqs
-            for f in 0..<freqs {
-                let idx = base + f
-                let rr = rVals[idx]
-                let ii = iVals[idx]
-                let mag = (rr * rr + ii * ii).squareRoot()
-                state[f] = mag * oneMinusA + state[f] * a
-                let den = state[f].squareRoot()
-                outR[idx] = rr / den
-                outI[idx] = ii / den
-            }
-        }
-
-        return (
-            MLXArray(outR).reshaped([frames, freqs]),
-            MLXArray(outI).reshaped([frames, freqs])
-        )
-    }
-
-    private func normAlpha() -> Float {
-        normAlphaValue
-    }
-
-    private static func computeNormAlpha(hopSize: Int, sampleRate: Int) -> Float {
-        let aRaw = exp(-Float(hopSize) / Float(sampleRate))
-        var precision = 3
-        var a: Float = 1.0
-        while a >= 1.0 {
-            let scale = powf(10, Float(precision))
-            a = (aRaw * scale).rounded() / scale
-            precision += 1
-        }
-        return a
-    }
-
-    private func applyLookahead(feature: MLXArray, lookahead: Int) -> MLXArray {
-        guard lookahead > 0 else { return feature }
-        let t = feature.shape[2]
-        guard t > lookahead else { return feature }
-        let shifted = feature[0..., 0..., lookahead..<t, 0...]
-        let pad = MLXArray.zeros([feature.shape[0], feature.shape[1], lookahead, feature.shape[3]], type: Float.self)
-        return MLX.concatenated([shifted, pad], axis: 2)
-    }
-
-    // MARK: - Utility
-
-    private func erbEnergies(_ specMagSq: MLXArray) -> MLXArray {
-        if erbFB.shape.count == 2,
-           erbFB.shape[0] == config.freqBins,
-           erbFB.shape[1] == config.nbErb
-        {
-            // Use learned ERB filterbank projection (same operation as streaming path).
-            return MLX.matmul(specMagSq, erbFB.asType(specMagSq.dtype))
-        }
-
-        var bands = [MLXArray]()
-        bands.reserveCapacity(erbBandWidths.count)
-        var start = 0
-        for width in erbBandWidths {
-            let stop = min(start + width, config.freqBins)
-            if stop > start {
-                bands.append(MLX.mean(specMagSq[0..., start..<stop], axis: 1))
-            } else {
-                bands.append(MLXArray.zeros([specMagSq.shape[0]], type: Float.self))
-            }
-            start = stop
-        }
-        return MLX.stacked(bands, axis: 1)
-    }
-
-    private static func libdfFreqToErb(_ freqHz: Float) -> Float {
-        9.265 * log1p(freqHz / (24.7 * 9.265))
-    }
-
-    private static func libdfErbToFreq(_ erb: Float) -> Float {
-        24.7 * 9.265 * (exp(erb / 9.265) - 1.0)
-    }
-
-    private static func libdfErbBandWidths(
-        sampleRate: Int,
-        fftSize: Int,
-        nbBands: Int,
-        minNbFreqs: Int
-    ) -> [Int] {
-        guard sampleRate > 0, fftSize > 0, nbBands > 0 else { return [] }
-
-        let nyq = sampleRate / 2
-        let freqWidth = Float(sampleRate) / Float(fftSize)
-        let erbLow = libdfFreqToErb(0)
-        let erbHigh = libdfFreqToErb(Float(nyq))
-        let step = (erbHigh - erbLow) / Float(nbBands)
-
-        var widths = [Int](repeating: 0, count: nbBands)
-        var prevFreq = 0
-        var freqOver = 0
-        let minBins = max(1, minNbFreqs)
-
-        for i in 1...nbBands {
-            let f = libdfErbToFreq(erbLow + Float(i) * step)
-            let fb = Int((f / freqWidth).rounded())
-            var nbFreqs = fb - prevFreq - freqOver
-            if nbFreqs < minBins {
-                freqOver = minBins - nbFreqs
-                nbFreqs = minBins
-            } else {
-                freqOver = 0
-            }
-            widths[i - 1] = max(1, nbFreqs)
-            prevFreq = fb
-        }
-
-        widths[nbBands - 1] += 1  // fft_size/2 + 1 bins
-        let target = fftSize / 2 + 1
-        let total = widths.reduce(0, +)
-        if total > target {
-            widths[nbBands - 1] -= (total - target)
-        } else if total < target {
-            widths[nbBands - 1] += (target - total)
-        }
-        return widths
-    }
-
-    private static func vorbisWindow(size: Int) -> MLXArray {
-        let half = max(1, size / 2)
-        var window = [Float](repeating: 0, count: size)
-        for i in 0..<size {
-            let inner = sin(0.5 * Float.pi * (Float(i) + 0.5) / Float(half))
-            window[i] = sin(0.5 * Float.pi * inner * inner)
-        }
-        return MLXArray(window)
-    }
-
-    private static func linspace(start: Float, end: Float, count: Int) -> [Float] {
-        guard count > 1 else { return [start] }
-        let step = (end - start) / Float(count - 1)
-        return (0..<count).map { start + Float($0) * step }
-    }
-
-    private static func buildBatchNormAffine(weights: [String: MLXArray]) -> ([String: MLXArray], [String: MLXArray]) {
-        var scaleByPrefix = [String: MLXArray]()
-        var biasByPrefix = [String: MLXArray]()
-
-        for (key, mean) in weights where key.hasSuffix(".running_mean") {
-            let prefix = String(key.dropLast(".running_mean".count))
-            guard let gamma = weights["\(prefix).weight"],
-                  let beta = weights["\(prefix).bias"],
-                  let variance = weights["\(prefix).running_var"]
-            else {
-                continue
-            }
-            let scale = gamma / MLX.sqrt(variance + MLXArray(Float(1e-5)))
-            let shift = beta - mean * scale
-            scaleByPrefix[prefix] = scale.reshaped([1, scale.shape[0], 1, 1])
-            biasByPrefix[prefix] = shift.reshaped([1, shift.shape[0], 1, 1])
-        }
-
-        return (scaleByPrefix, biasByPrefix)
-    }
-
-    private static func buildConv2dWeightCache(weights: [String: MLXArray]) -> [String: MLXArray] {
-        var cache = [String: MLXArray]()
-        for (key, weight) in weights where key.hasSuffix(".weight") && weight.ndim == 4 {
-            cache[key] = weight.transposed(0, 2, 3, 1)
-        }
-        return cache
-    }
-
-    private static func buildGRUTransposedWeightCache(weights: [String: MLXArray]) -> [String: MLXArray] {
-        var cache = [String: MLXArray]()
-        cache.reserveCapacity(24)
-        for (key, weight) in weights where key.contains(".gru.weight_") && weight.ndim == 2 {
-            cache[key] = weight.transposed()
-        }
-        return cache
-    }
-
-    private static func buildV1GroupedLinearPacks(
-        weights: [String: MLXArray],
-        groups: Int
-    ) -> [String: V1GroupedLinearPack] {
-        guard groups > 1 else { return [:] }
-        var groupedKeys = [String: [Int]]()
-        for key in weights.keys where key.hasSuffix(".weight") {
-            let parts = key.split(separator: ".")
-            guard parts.count >= 3, let idx = Int(parts[parts.count - 2]) else { continue }
-            let prefix = parts.dropLast(2).joined(separator: ".")
-            groupedKeys[prefix, default: []].append(idx)
-        }
-
-        var packs = [String: V1GroupedLinearPack]()
-        for (prefix, idxs) in groupedKeys {
-            let unique = Array(Set(idxs)).sorted()
-            guard unique.count == groups, unique.first == 0, unique.last == groups - 1 else { continue }
-
-            var weightSlices = [MLXArray]()
-            var biasSlices = [MLXArray]()
-            weightSlices.reserveCapacity(groups)
-            biasSlices.reserveCapacity(groups)
-            var valid = true
-            for g in 0..<groups {
-                guard let wg = weights["\(prefix).\(g).weight"],
-                      let bg = weights["\(prefix).\(g).bias"],
-                      wg.ndim == 2, bg.ndim == 1
-                else {
-                    valid = false
-                    break
-                }
-                weightSlices.append(wg.transposed())  // [I, O]
-                biasSlices.append(bg)
-            }
-            guard valid else { continue }
-
-            let weightGIO = MLX.stacked(weightSlices, axis: 0)
-            let biasGO = MLX.stacked(biasSlices, axis: 0)
-            packs[prefix] = V1GroupedLinearPack(
-                weightGIO: weightGIO,
-                biasGO: biasGO,
-                groups: groups,
-                inputPerGroup: weightGIO.shape[1],
-                outputPerGroup: weightGIO.shape[2]
-            )
-        }
-        return packs
-    }
-
-    private static func buildV1GroupedGRUPacks(
-        weights: [String: MLXArray],
-        groups: Int,
-        prefixes: [String]
-    ) -> [String: V1GroupedGRUPack] {
-        guard groups > 1 else { return [:] }
-        var packs = [String: V1GroupedGRUPack]()
-
-        for prefix in prefixes {
-            var layers = [V1GroupedGRULayerPack]()
-            var layerIndex = 0
-            while true {
-                var wih = [MLXArray]()
-                var whh = [MLXArray]()
-                var bih = [MLXArray]()
-                var bhh = [MLXArray]()
-                wih.reserveCapacity(groups)
-                whh.reserveCapacity(groups)
-                bih.reserveCapacity(groups)
-                bhh.reserveCapacity(groups)
-
-                var hasLayer = true
-                for g in 0..<groups {
-                    let base = "\(prefix).\(layerIndex).layers.\(g)"
-                    guard let wihG = weights["\(base).weight_ih_l0"],
-                          let whhG = weights["\(base).weight_hh_l0"],
-                          let bihG = weights["\(base).bias_ih_l0"],
-                          let bhhG = weights["\(base).bias_hh_l0"],
-                          wihG.ndim == 2, whhG.ndim == 2, bihG.ndim == 1, bhhG.ndim == 1
-                    else {
-                        hasLayer = false
-                        break
-                    }
-                    wih.append(wihG.transposed())  // [I, 3H]
-                    whh.append(whhG.transposed())  // [H, 3H]
-                    bih.append(bihG)
-                    bhh.append(bhhG)
-                }
-                guard hasLayer else { break }
-
-                let wihGI3H = MLX.stacked(wih, axis: 0)
-                let whhGH3H = MLX.stacked(whh, axis: 0)
-                let bihG3H = MLX.stacked(bih, axis: 0)
-                let bhhG3H = MLX.stacked(bhh, axis: 0)
-                layers.append(
-                    V1GroupedGRULayerPack(
-                        weightIHGI3H: wihGI3H,
-                        weightHHGH3H: whhGH3H,
-                        biasIHG3H: bihG3H,
-                        biasHHG3H: bhhG3H,
-                        inputPerGroup: wihGI3H.shape[1],
-                        hiddenPerGroup: wihGI3H.shape[2] / 3
-                    )
-                )
-                layerIndex += 1
-            }
-
-            if !layers.isEmpty {
-                packs[prefix] = V1GroupedGRUPack(groups: groups, layers: layers)
-            }
-        }
-
-        return packs
-    }
-
-    private static func buildGroupedTransposeWeights(
-        weights: [String: MLXArray],
-        groups: Int
-    ) -> [String: [MLXArray]] {
-        guard groups > 1 else { return [:] }
-        var cache = [String: [MLXArray]]()
-        for (key, weight) in weights where key.hasSuffix(".0.weight") && weight.ndim == 4 {
-            guard weight.shape[0] % groups == 0 else { continue }
-            let inPerGroup = weight.shape[0] / groups
-            var grouped = [MLXArray]()
-            grouped.reserveCapacity(groups)
-            for g in 0..<groups {
-                let inStart = g * inPerGroup
-                let inEnd = inStart + inPerGroup
-                let wg = weight[inStart..<inEnd, 0..., 0..., 0...]
-                grouped.append(wg.transposed(1, 2, 3, 0))
-            }
-            cache[key] = grouped
-        }
-        return cache
-    }
-
-    private static func buildDenseTransposeWeights(
-        weights: [String: MLXArray],
-        groups: Int
-    ) -> [String: MLXArray] {
-        guard groups > 1 else { return [:] }
-        var cache = [String: MLXArray]()
-        for (key, weight) in weights where key.hasSuffix(".0.weight") && weight.ndim == 4 {
-            guard weight.shape[0] % groups == 0 else { continue }
-            let inPerGroup = weight.shape[0] / groups
-            let outPerGroup = weight.shape[1]
-            let kT = weight.shape[2]
-            let kF = weight.shape[3]
-            let totalIn = inPerGroup * groups
-
-            var outBlocks = [MLXArray]()
-            outBlocks.reserveCapacity(groups)
-            for g in 0..<groups {
-                let inStart = g * inPerGroup
-                let inEnd = inStart + inPerGroup
-                let wg = weight[inStart..<inEnd, 0..., 0..., 0...].transposed(1, 2, 3, 0)  // [out_pg, kT, kF, in_pg]
-                let leftChannels = g * inPerGroup
-                let rightChannels = totalIn - leftChannels - inPerGroup
-                let left = MLXArray.zeros([outPerGroup, kT, kF, leftChannels], type: Float.self)
-                let right = MLXArray.zeros([outPerGroup, kT, kF, rightChannels], type: Float.self)
-                outBlocks.append(MLX.concatenated([left, wg, right], axis: 3))
-            }
-            cache[key] = MLX.concatenated(outBlocks, axis: 0)  // [groups*out_pg, kT, kF, groups*in_pg]
-        }
-        return cache
-    }
-
-    private func w(_ key: String) throws -> MLXArray {
-        guard let value = weights[key] else {
-            throw DeepFilterNetError.missingWeightKey(key)
-        }
-        return value
     }
 }
